refactor(commands): split ads.py into modular components (#603)
Breaks the 3 156-line direct_cli/commands/ads.py into a command package
with type-specific modules, keeping the CLI surface byte-identical.

Layout:
- ads/_cli.py — thin Click router (group + get/add/update + lifecycle);
- ads/objects.py — build_ad_object / build_ad_update_object (pure builders);
- ads/batch.py — --from-file / --ads-json row normalization + chunked send;
- ads/base.py — shared validation/parsing helpers;
- ads/text.py, ads/responsive.py, ads/shopping.py, ads/mobile_app.py,
  ads/builder.py — per-subtype payload assembly;
- ads/__init__.py — re-exports ads, build_ad_object, build_ad_update_object.

No CLI surface changes (commands, flags, payloads unchanged). The
create_client patchability contract is preserved: the package exposes
create_client (re-exported from _cli) and scripts/build_api_coverage_report.py
now patches the _cli submodule when a group is a package.
