--- direct_cli/commands/ads.py
@@ -1,3145 +1,0 @@
-"""
-Ads commands
-"""
-
-from typing import Any, Dict, Optional
-
-import click
-
-from ..api import client_from_ctx, create_client
-from ..i18n import t
-from ..output import format_output, handle_api_errors
-from ._execute import execute_request
-from ._lifecycle import register_lifecycle_commands
-from . import _batch
-from ..utils import (
-    add_criteria_csv,
-    build_common_params,
-    enforce_criteria_array_limits,
-    get_default_fields,
-    MICRO_RUBLES,
-    parse_condition_specs,
-    parse_csv_strings,
-    parse_ids,
-    parse_nested_field_names,
-)
-
-# Yandex Direct ads.get caps SelectionCriteria arrays at runtime (the WSDL
-# declares them maxOccurs="unbounded"). Live measurement 2026-06-17 via sandbox:
-# --campaign-ids ×11 → 4001 "Exceed the maximum number of IDs per array
-# SelectionCriteria.CampaignIds"; --adgroup-ids ×10001 → 4001 ".AdGroupIds"
-# (N=1000 accepted); --vcard-ids ×11 (with anchor --campaign-ids 1) → 4001
-# ".VCardIds"; --sitelink-set-ids ×11 (with anchor) → 4001 ".SitelinkSetIds".
-# Ids and AdExtensionIds accepted at N=10000.
-ADS_GET_CRITERIA_LIMITS = {
-    "CampaignIds": 10,
-    "AdGroupIds": 1000,
-    "VCardIds": 10,
-    "SitelinkSetIds": 10,
-}
-
-
-@click.group()
-def ads():
-    """Manage ads"""
-
-
-MOBILE_APP_FEATURES = ("PRICE", "ICON", "CUSTOMER_RATING", "RATINGS")
-
-FEED_BASED_UPDATE_FIELDS = {
-    "sitelink_set_id",
-    "callouts_add",
-    "callouts_remove",
-    "callouts_set",
-    "business_id",
-    "feed_filter_conditions",
-    "title_sources",
-    "text_sources",
-    "default_texts",
-}
-
-FEED_BASED_ADD_FIELDS = {
-    "sitelink_set_id",
-    "ad_extensions",
-    "business_id",
-    "feed_id",
-    "feed_filter_conditions",
-    "title_sources",
-    "text_sources",
-    "default_texts",
-}
-
-DYNAMIC_TEXT_AD_ADD_FIELDS = {
-    "text",
-    "image_hash",
-    "vcard_id",
-    "sitelink_set_id",
-    "ad_extensions",
-}
-
-MOBILE_APP_AD_ADD_FIELDS = {
-    "title",
-    "text",
-    "image_hash",
-    "action",
-    "tracking_url",
-    "age_label",
-    "mobile_app_features",
-    "video_extension_creative_id",
-    "erir_ad_description",
-}
-
-MOBILE_APP_IMAGE_AD_ADD_FIELDS = {
-    "image_hash",
-    "tracking_url",
-    "erir_ad_description",
-}
-
-SMART_AD_BUILDER_ADD_FIELDS = {"logo_extension_hash"}
-
-
-TEXT_AD_UPDATE_FIELDS = {
-    "title",
-    "text",
-    "href",
-    "image_hash",
-    "clear_image_hash",
-    "title2",
-    "display_url_path",
-    "vcard_id",
-    "sitelink_set_id",
-    "turbo_page_id",
-    "callouts_add",
-    "callouts_remove",
-    "callouts_set",
-    "video_extension_creative_id",
-    "price_extension_price",
-    "price_extension_old_price",
-    "price_extension_price_qualifier",
-    "price_extension_price_currency",
-    "final_url",
-    "age_label",
-    "business_id",
-    "prefer_vcard_over_business",
-    "erir_ad_description",
-}
-
-MOBILE_APP_AD_UPDATE_FIELDS = {
-    "title",
-    "text",
-    "image_hash",
-    "clear_image_hash",
-    "action",
-    "tracking_url",
-    "age_label",
-    "mobile_app_features",
-    "video_extension_creative_id",
-    "erir_ad_description",
-}
-
-# NOTE: TEXT_IMAGE_AD / MOBILE_APP_IMAGE_AD share WSDL ImageAdUpdateBase, whose
-# AdImageHash is NOT nillable (unlike TextAdUpdateBase / MobileAppAdBase). The
-# live API rejects ``AdImageHash: null`` for these two subtypes with error 8000
-# ("AdImageHash cannot have the null value"), so --clear-image-hash is
-# deliberately absent here — the parity gate rejects it as an incompatible flag.
-TEXT_IMAGE_AD_UPDATE_FIELDS = {
-    "image_hash",
-    "final_url",
-    "href",
-    "turbo_page_id",
-    "erir_ad_description",
-}
-
-MOBILE_APP_IMAGE_UPDATE_FIELDS = {
-    "image_hash",
-    "tracking_url",
-    "erir_ad_description",
-}
-
-SMART_AD_BUILDER_UPDATE_FIELDS = {
-    "logo_extension_hash",
-    "erir_ad_description",
-}
-
-AD_BUILDER_BASE_UPDATE_FIELDS = {
-    "creative_id",
-    "creative_erir_ad_description",
-    "erir_ad_description",
-}
-
-AD_BUILDER_TYPE_FIELDS = {
-    "TEXT_AD_BUILDER_AD": AD_BUILDER_BASE_UPDATE_FIELDS
-    | {"final_url", "href", "turbo_page_id"},
-    "MOBILE_APP_AD_BUILDER_AD": AD_BUILDER_BASE_UPDATE_FIELDS | {"tracking_url"},
-    "MOBILE_APP_CPC_VIDEO_AD_BUILDER_AD": AD_BUILDER_BASE_UPDATE_FIELDS
-    | {"tracking_url"},
-    "CPC_VIDEO_AD_BUILDER_AD": AD_BUILDER_BASE_UPDATE_FIELDS
-    | {"href", "turbo_page_id"},
-    "CPM_BANNER_AD_BUILDER_AD": AD_BUILDER_BASE_UPDATE_FIELDS
-    | {"href", "tracking_pixels", "turbo_page_id"},
-    "CPM_VIDEO_AD_BUILDER_AD": AD_BUILDER_BASE_UPDATE_FIELDS
-    | {"href", "tracking_pixels", "turbo_page_id"},
-}
-
-AD_BUILDER_UPDATE_BLOCKS = {
-    "TEXT_AD_BUILDER_AD": "TextAdBuilderAd",
-    "MOBILE_APP_AD_BUILDER_AD": "MobileAppAdBuilderAd",
-    "MOBILE_APP_CPC_VIDEO_AD_BUILDER_AD": "MobileAppCpcVideoAdBuilderAd",
-    "CPC_VIDEO_AD_BUILDER_AD": "CpcVideoAdBuilderAd",
-    "CPM_BANNER_AD_BUILDER_AD": "CpmBannerAdBuilderAd",
-    "CPM_VIDEO_AD_BUILDER_AD": "CpmVideoAdBuilderAd",
-}
-
-AD_BUILDER_ADD_BLOCKS = {
-    "TEXT_AD_BUILDER_AD": "TextAdBuilderAd",
-    "MOBILE_APP_AD_BUILDER_AD": "MobileAppAdBuilderAd",
-    "MOBILE_APP_CPC_VIDEO_AD_BUILDER_AD": "MobileAppCpcVideoAdBuilderAd",
-    "CPC_VIDEO_AD_BUILDER_AD": "CpcVideoAdBuilderAd",
-    "CPM_BANNER_AD_BUILDER_AD": "CpmBannerAdBuilderAd",
-    "CPM_VIDEO_AD_BUILDER_AD": "CpmVideoAdBuilderAd",
-}
-
-AD_BUILDER_ADD_BASE_FIELDS = {"creative_id", "erir_ad_description"}
-
-AD_BUILDER_ADD_TYPE_FIELDS = {
-    "TEXT_AD_BUILDER_AD": AD_BUILDER_ADD_BASE_FIELDS
-    | {"final_url", "href", "turbo_page_id"},
-    "MOBILE_APP_AD_BUILDER_AD": AD_BUILDER_ADD_BASE_FIELDS | {"tracking_url"},
-    "MOBILE_APP_CPC_VIDEO_AD_BUILDER_AD": AD_BUILDER_ADD_BASE_FIELDS | {"tracking_url"},
-    "CPC_VIDEO_AD_BUILDER_AD": AD_BUILDER_ADD_BASE_FIELDS | {"href", "turbo_page_id"},
-    "CPM_BANNER_AD_BUILDER_AD": AD_BUILDER_ADD_BASE_FIELDS
-    | {"href", "tracking_pixels", "turbo_page_id"},
-    "CPM_VIDEO_AD_BUILDER_AD": AD_BUILDER_ADD_BASE_FIELDS
-    | {"href", "tracking_pixels", "turbo_page_id"},
-}
-
-AD_BUILDER_ADD_DESTINATION_TYPES = {
-    "TEXT_AD_BUILDER_AD",
-    "CPC_VIDEO_AD_BUILDER_AD",
-    "CPM_BANNER_AD_BUILDER_AD",
-    "CPM_VIDEO_AD_BUILDER_AD",
-}
-
-
-def _reject_incompatible_flags(
-    ad_type: str,
-    allowed_fields: set[str],
-    provided: dict[str, object],
-    flag_for: dict[str, str],
-) -> None:
-    """Reject typed flags that do not belong to the selected ad subtype."""
-    incompatible = [
-        flag_for[name]
-        for name, value in provided.items()
-        if value is not None and value != () and name not in allowed_fields
-    ]
-    if incompatible:
-        allowed_flags = ", ".join(sorted(flag_for[name] for name in allowed_fields))
-        raise click.UsageError(
-            t(
-                "{arg0} is not compatible with --type {ad_type}. Allowed flags for {ad_type}: {allowed_flags}."
-            ).format(
-                arg0=", ".join(incompatible),
-                ad_type=ad_type,
-                allowed_flags=allowed_flags,
-            )
-        )
-
-
-def _build_callout_setting(callouts_add, callouts_remove, callouts_set):
-    """Build AdExtensionSetting for text-like ad update payloads.
-
-    SET is mutually exclusive with ADD/REMOVE per WSDL OperationEnum
-    semantics. Returns None when no callout flag was provided.
-    """
-    if callouts_set and (callouts_add or callouts_remove):
-        raise click.UsageError(
-            t(
-                "--callouts-set is mutually exclusive with "
-                "--callouts-add / --callouts-remove. "
-                "Use --callouts-set to replace the full callout list, "
-                "or --callouts-add / --callouts-remove for incremental edits."
-            )
-        )
-    items = []
-    for csv_value, op in (
-        (callouts_set, "SET"),
-        (callouts_add, "ADD"),
-        (callouts_remove, "REMOVE"),
-    ):
-        if csv_value is None:
-            continue
-        try:
-            ids = parse_ids(csv_value)
-        except ValueError as exc:
-            raise click.UsageError(
-                t("--callouts-{arg0}: {exc}").format(arg0=op.lower(), exc=exc)
-            )
-        if not ids:
-            raise click.UsageError(
-                t(
-                    "--callouts-{arg0} must contain at least one ad extension ID."
-                ).format(arg0=op.lower())
-            )
-        for ad_ext_id in ids:
-            items.append({"AdExtensionId": ad_ext_id, "Operation": op})
-    if not items:
-        return None
-    return {"AdExtensions": items}
-
-
-def _build_text_ad_update_base(
-    vcard_id: Optional[int],
-    image_hash: Optional[str],
-    sitelink_set_id: Optional[int],
-    callout_setting: Optional[dict[str, Any]],
-    clear_image_hash: bool = False,
-) -> dict[str, object]:
-    """Build fields inherited from WSDL TextAdUpdateBase."""
-    text_ad_base: dict[str, object] = {}
-    if vcard_id is not None:
-        text_ad_base["VCardId"] = vcard_id
-    if clear_image_hash:
-        text_ad_base["AdImageHash"] = None
-    elif image_hash:
-        text_ad_base["AdImageHash"] = image_hash
-    if sitelink_set_id is not None:
-        text_ad_base["SitelinkSetId"] = sitelink_set_id
-    if callout_setting:
-        text_ad_base["CalloutSetting"] = callout_setting
-    return text_ad_base
-
-
-def _build_price_extension(
-    price_extension_price,
-    price_extension_old_price,
-    price_extension_price_qualifier,
-    price_extension_price_currency,
-):
-    """Build TextAd.PriceExtension update payload from typed flags."""
-    price_extension = {}
-    if price_extension_price is not None:
-        price_extension["Price"] = price_extension_price
-    if price_extension_old_price is not None:
-        price_extension["OldPrice"] = price_extension_old_price
-    if price_extension_price_qualifier:
-        price_extension["PriceQualifier"] = price_extension_price_qualifier.upper()
-    if price_extension_price_currency:
-        price_extension["PriceCurrency"] = price_extension_price_currency.upper()
-    return price_extension or None
-
-
-def _build_price_extension_add(
-    price_extension_price,
-    price_extension_old_price,
-    price_extension_price_qualifier,
-    price_extension_price_currency,
-    container_name="TextAd",
-):
-    """Build PriceExtension add payload for an ad subtype from typed flags."""
-    provided = (
-        price_extension_price is not None
-        or price_extension_old_price is not None
-        or price_extension_price_qualifier
-        or price_extension_price_currency
-    )
-    if not provided:
-        return None
-
-    missing = []
-    if price_extension_price is None:
-        missing.append("--price-extension-price")
-    if not price_extension_price_qualifier:
-        missing.append("--price-extension-price-qualifier")
-    if not price_extension_price_currency:
-        missing.append("--price-extension-price-currency")
-    if missing:
-        raise click.UsageError(
-            t("{container_name}.PriceExtension add requires {arg0}").format(
-                container_name=container_name, arg0=", ".join(missing)
-            )
-        )
-
-    price_extension = {
-        "Price": price_extension_price,
-        "PriceQualifier": price_extension_price_qualifier.upper(),
-        "PriceCurrency": price_extension_price_currency.upper(),
-    }
-    if price_extension_old_price is not None:
-        price_extension["OldPrice"] = price_extension_old_price
-    return price_extension
-
-
-def _parse_required_csv_strings(
-    csv_value: Optional[str], flag_name: str
-) -> Optional[list[str]]:
-    """Parse a comma-separated string list that must not be empty if present."""
-    if csv_value is None:
-        return None
-    values = parse_csv_strings(csv_value)
-    if not values:
-        raise click.UsageError(
-            t("{flag_name} must contain at least one value.").format(
-                flag_name=flag_name
-            )
-        )
-    return values
-
-
-def _parse_required_ids(
-    csv_value: Optional[str], flag_name: str
-) -> Optional[list[int]]:
-    """Parse a comma-separated integer list that must not be empty if present."""
-    if csv_value is None:
-        return None
-    try:
-        ids = parse_ids(csv_value)
-    except ValueError as exc:
-        raise click.UsageError(
-            t("{flag_name}: {exc}").format(flag_name=flag_name, exc=exc)
-        )
-    if not ids:
-        raise click.UsageError(
-            t("{flag_name} must contain at least one ID.").format(flag_name=flag_name)
-        )
-    return ids
-
-
-def _parse_mobile_app_features(
-    raw_values: tuple[str, ...],
-) -> Optional[list[dict[str, str]]]:
-    """Parse repeatable MobileAppAd.Features items as FEATURE=YES|NO."""
-    if not raw_values:
-        return None
-
-    items: list[dict[str, str]] = []
-    allowed_features = ", ".join(MOBILE_APP_FEATURES)
-    for raw_value in raw_values:
-        feature_raw, separator, enabled_raw = raw_value.strip().partition("=")
-        if not separator:
-            raise click.UsageError(
-                t(
-                    "--mobile-app-feature expects FEATURE=YES|NO "
-                    "(for example PRICE=YES)."
-                )
-            )
-
-        feature = feature_raw.strip().upper()
-        enabled = enabled_raw.strip().upper()
-        if feature not in MOBILE_APP_FEATURES:
-            raise click.UsageError(
-                t(
-                    "Invalid --mobile-app-feature feature {feature_raw!r}; allowed: {allowed_features}."
-                ).format(feature_raw=feature_raw, allowed_features=allowed_features)
-            )
-        if enabled not in {"YES", "NO"}:
-            raise click.UsageError(
-                t(
-                    "Invalid --mobile-app-feature value {enabled_raw!r}; expected YES or NO."
-                ).format(enabled_raw=enabled_raw)
-            )
-
-        items.append({"Feature": feature, "Enabled": enabled})
-
-    return items
-
-
-def _build_responsive_ad_update(
-    texts: Optional[str],
-    titles: Optional[str],
-    image_hashes: Optional[str],
-    video_extension_ids: Optional[str],
-    sitelink_set_id: Optional[int],
-    callout_setting: Optional[dict[str, Any]],
-    href: Optional[str],
-    age_label: Optional[str],
-    display_url_path: Optional[str],
-    price_extension: Optional[dict[str, object]],
-    business_id: Optional[int],
-    erir_ad_description: Optional[str],
-) -> dict[str, object]:
-    """Build ResponsiveAdUpdate payload from typed flags."""
-    responsive_ad: dict[str, object] = {}
-
-    parsed_texts = _parse_required_csv_strings(texts, "--texts")
-    if parsed_texts:
-        responsive_ad["Texts"] = parsed_texts
-    parsed_titles = _parse_required_csv_strings(titles, "--titles")
-    if parsed_titles:
-        responsive_ad["Titles"] = parsed_titles
-    parsed_image_hashes = _parse_required_csv_strings(image_hashes, "--image-hashes")
-    if parsed_image_hashes:
-        responsive_ad["AdImageHashes"] = {"Items": parsed_image_hashes}
-    parsed_video_extension_ids = _parse_required_ids(
-        video_extension_ids, "--video-extension-ids"
-    )
-    if parsed_video_extension_ids:
-        responsive_ad["VideoExtensionIds"] = {"Items": parsed_video_extension_ids}
-    if sitelink_set_id is not None:
-        responsive_ad["SitelinkSetId"] = sitelink_set_id
-    if callout_setting:
-        responsive_ad["CalloutSetting"] = callout_setting
-    if href:
-        responsive_ad["Href"] = href
-    if age_label:
-        responsive_ad["AgeLabel"] = age_label.upper()
-    if display_url_path:
-        responsive_ad["DisplayUrlPath"] = display_url_path
-    if price_extension:
-        responsive_ad["PriceExtension"] = price_extension
-    if business_id is not None:
-        responsive_ad["BusinessId"] = business_id
-    if erir_ad_description:
-        responsive_ad["ErirAdDescription"] = erir_ad_description
-
-    return responsive_ad
-
-
-def _build_feed_based_ad_update(
-    sitelink_set_id: Optional[int],
-    callout_setting: Optional[dict[str, Any]],
-    business_id: Optional[int],
-    feed_filter_conditions: tuple[str, ...],
-    title_sources: Optional[str],
-    text_sources: Optional[str],
-    default_texts: Optional[str],
-) -> dict[str, object]:
-    """Build ShoppingAdUpdate / ListingAdUpdate payload from typed flags."""
-    ad_payload: dict[str, object] = {}
-
-    if sitelink_set_id is not None:
-        ad_payload["SitelinkSetId"] = sitelink_set_id
-    if callout_setting:
-        ad_payload["CalloutSetting"] = callout_setting
-    if business_id is not None:
-        ad_payload["BusinessId"] = business_id
-    if feed_filter_conditions:
-        try:
-            parsed_conditions = parse_condition_specs(list(feed_filter_conditions))
-        except ValueError as exc:
-            raise click.UsageError(t("--feed-filter-condition: {exc}").format(exc=exc))
-        if parsed_conditions:
-            ad_payload["FeedFilterConditions"] = {"Items": parsed_conditions}
-
-    parsed_title_sources = _parse_required_csv_strings(title_sources, "--title-sources")
-    if parsed_title_sources:
-        ad_payload["TitleSources"] = {"Items": parsed_title_sources}
-    parsed_text_sources = _parse_required_csv_strings(text_sources, "--text-sources")
-    if parsed_text_sources:
-        ad_payload["TextSources"] = {"Items": parsed_text_sources}
-    parsed_default_texts = _parse_required_csv_strings(default_texts, "--default-texts")
-    if parsed_default_texts:
-        ad_payload["DefaultTexts"] = parsed_default_texts
-
-    return ad_payload
-
-
-def _build_feed_based_ad_add(
-    feed_id: Optional[int],
-    default_texts: Optional[str],
-    sitelink_set_id: Optional[int],
-    ad_extensions: Optional[str],
-    business_id: Optional[int],
-    feed_filter_conditions: tuple[str, ...],
-    title_sources: Optional[str],
-    text_sources: Optional[str],
-    container_name: str,
-) -> dict[str, object]:
-    """Build ShoppingAdAdd / ListingAdAdd payload from typed flags."""
-    missing_fields = []
-    if feed_id is None:
-        missing_fields.append("--feed-id")
-    if default_texts is None:
-        missing_fields.append("--default-texts")
-    if missing_fields:
-        raise click.UsageError(
-            t("{container_name} requires {arg0}").format(
-                container_name=container_name, arg0=", ".join(missing_fields)
-            )
-        )
-
-    default_text = default_texts.strip() if default_texts else ""
-    if not default_text:
-        raise click.UsageError(t("--default-texts must contain a value."))
-
-    ad_payload: dict[str, object] = {
-        "FeedId": feed_id,
-        "DefaultTexts": [default_text],
-    }
-
-    if sitelink_set_id is not None:
-        ad_payload["SitelinkSetId"] = sitelink_set_id
-    parsed_ad_extensions = _parse_required_ids(ad_extensions, "--ad-extensions")
-    if parsed_ad_extensions:
-        ad_payload["AdExtensionIds"] = parsed_ad_extensions
-    if business_id is not None:
-        ad_payload["BusinessId"] = business_id
-    if feed_filter_conditions:
-        if len(feed_filter_conditions) > 30:
-            raise click.UsageError(
-                t(
-                    "{container_name}.FeedFilterConditions accepts at most 30 filters."
-                ).format(container_name=container_name)
-            )
-        try:
-            parsed_conditions = parse_condition_specs(list(feed_filter_conditions))
-        except ValueError as exc:
-            raise click.UsageError(t("--feed-filter-condition: {exc}").format(exc=exc))
-        if parsed_conditions:
-            ad_payload["FeedFilterConditions"] = parsed_conditions
-
-    parsed_title_sources = _parse_required_csv_strings(title_sources, "--title-sources")
-    if parsed_title_sources:
-        ad_payload["TitleSources"] = parsed_title_sources
-    parsed_text_sources = _parse_required_csv_strings(text_sources, "--text-sources")
-    if parsed_text_sources:
-        ad_payload["TextSources"] = parsed_text_sources
-
-    return ad_payload
-
-
-def _build_dynamic_text_ad_add(
-    text: Optional[str],
-    image_hash: Optional[str],
-    vcard_id: Optional[int],
-    sitelink_set_id: Optional[int],
-    ad_extensions: Optional[str],
-) -> dict[str, object]:
-    """Build DynamicTextAdAdd payload from typed flags."""
-    if not text:
-        raise click.UsageError(t("DYNAMIC_TEXT_AD requires --text"))
-
-    dynamic_text_ad: dict[str, object] = {"Text": text}
-    if image_hash:
-        dynamic_text_ad["AdImageHash"] = image_hash
-    if vcard_id is not None:
-        dynamic_text_ad["VCardId"] = vcard_id
-    if sitelink_set_id is not None:
-        dynamic_text_ad["SitelinkSetId"] = sitelink_set_id
-    parsed_ad_extensions = _parse_required_ids(ad_extensions, "--ad-extensions")
-    if parsed_ad_extensions:
-        dynamic_text_ad["AdExtensionIds"] = parsed_ad_extensions
-
-    return dynamic_text_ad
-
-
-def _build_ad_builder_update(
-    creative_id: Optional[int],
-    creative_erir_ad_description: Optional[str],
-    erir_ad_description: Optional[str],
-    final_url: Optional[str],
-    href: Optional[str],
-    turbo_page_id: Optional[int],
-    tracking_url: Optional[str],
-    tracking_pixels: Optional[str],
-) -> dict[str, object]:
-    """Build an AdBuilder*Update payload from typed flags."""
-    ad_payload: dict[str, object] = {}
-
-    if creative_erir_ad_description and creative_id is None:
-        raise click.UsageError(
-            t("--creative-erir-ad-description requires --creative-id.")
-        )
-    if creative_id is not None:
-        creative: dict[str, object] = {"CreativeId": creative_id}
-        if creative_erir_ad_description:
-            creative["ErirAdDescription"] = creative_erir_ad_description
-        ad_payload["Creative"] = creative
-
-    if erir_ad_description:
-        ad_payload["ErirAdDescription"] = erir_ad_description
-    if final_url:
-        ad_payload["FinalUrl"] = final_url
-    if href:
-        ad_payload["Href"] = href
-    if turbo_page_id is not None:
-        ad_payload["TurboPageId"] = turbo_page_id
-    if tracking_url:
-        ad_payload["TrackingUrl"] = tracking_url
-
-    parsed_tracking_pixels = _parse_required_csv_strings(
-        tracking_pixels, "--tracking-pixels"
-    )
-    if parsed_tracking_pixels:
-        ad_payload["TrackingPixels"] = {"Items": parsed_tracking_pixels}
-
-    return ad_payload
-
-
-def _build_ad_builder_add(
-    creative_id: Optional[int],
-    erir_ad_description: Optional[str],
-    final_url: Optional[str],
-    href: Optional[str],
-    turbo_page_id: Optional[int],
-    tracking_url: Optional[str],
-    tracking_pixels: Optional[str],
-    ad_type: str,
-    container_name: str,
-) -> dict[str, object]:
-    """Build an AdBuilder*Add payload from typed flags."""
-    if creative_id is None:
-        raise click.UsageError(
-            t("{container_name} requires --creative-id.").format(
-                container_name=container_name
-            )
-        )
-    if (
-        ad_type in AD_BUILDER_ADD_DESTINATION_TYPES
-        and not href
-        and turbo_page_id is None
-    ):
-        raise click.UsageError(
-            t("{container_name} requires either --href or --turbo-page-id.").format(
-                container_name=container_name
-            )
-        )
-
-    ad_payload: dict[str, object] = {"Creative": {"CreativeId": creative_id}}
-
-    if erir_ad_description:
-        ad_payload["ErirAdDescription"] = erir_ad_description
-    if final_url:
-        ad_payload["FinalUrl"] = final_url
-    if href:
-        ad_payload["Href"] = href
-    if turbo_page_id is not None:
-        ad_payload["TurboPageId"] = turbo_page_id
-    if tracking_url:
-        ad_payload["TrackingUrl"] = tracking_url
-
-    parsed_tracking_pixels = _parse_required_csv_strings(
-        tracking_pixels, "--tracking-pixels"
-    )
-    if parsed_tracking_pixels:
-        ad_payload["TrackingPixels"] = {"Items": parsed_tracking_pixels}
-
-    return ad_payload
-
-
-def _build_mobile_app_image_ad_update(
-    image_hash: Optional[str],
-    erir_ad_description: Optional[str],
-    tracking_url: Optional[str],
-) -> dict[str, object]:
-    """Build MobileAppImageAdUpdate payload from typed flags.
-
-    No ``clear_image_hash`` path: ImageAdUpdateBase.AdImageHash is not nillable
-    and the live API rejects ``null`` for this subtype (error 8000).
-    """
-    mobile_app_image_ad: dict[str, object] = {}
-
-    if image_hash:
-        mobile_app_image_ad["AdImageHash"] = image_hash
-    if erir_ad_description:
-        mobile_app_image_ad["ErirAdDescription"] = erir_ad_description
-    if tracking_url:
-        mobile_app_image_ad["TrackingUrl"] = tracking_url
-
-    return mobile_app_image_ad
-
-
-def _build_mobile_app_image_ad_add(
-    image_hash: Optional[str],
-    erir_ad_description: Optional[str],
-    tracking_url: Optional[str],
-) -> dict[str, object]:
-    """Build MobileAppImageAdAdd payload from typed flags."""
-    if not image_hash:
-        raise click.UsageError(t("MOBILE_APP_IMAGE_AD requires --image-hash"))
-
-    mobile_app_image_ad: dict[str, object] = {"AdImageHash": image_hash}
-    if erir_ad_description:
-        mobile_app_image_ad["ErirAdDescription"] = erir_ad_description
-    if tracking_url:
-        mobile_app_image_ad["TrackingUrl"] = tracking_url
-
-    return mobile_app_image_ad
-
-
-def _build_smart_ad_builder_ad_add(
-    logo_extension_hash: Optional[str],
-) -> dict[str, object]:
-    """Build SmartAdBuilderAdAdd payload from typed flags."""
-    smart_ad_builder_ad: dict[str, object] = {}
-
-    if logo_extension_hash:
-        smart_ad_builder_ad["LogoExtensionHash"] = logo_extension_hash
-
-    return smart_ad_builder_ad
-
-
-def _build_smart_ad_builder_ad_update(
-    logo_extension_hash: Optional[str],
-    erir_ad_description: Optional[str],
-) -> dict[str, object]:
-    """Build SmartAdBuilderAdUpdate payload from typed flags."""
-    smart_ad_builder_ad: dict[str, object] = {}
-
-    if logo_extension_hash:
-        smart_ad_builder_ad["LogoExtensionHash"] = logo_extension_hash
-    if erir_ad_description:
-        smart_ad_builder_ad["ErirAdDescription"] = erir_ad_description
-
-    return smart_ad_builder_ad
-
-
-@ads.command()
-@click.option("--ids", help="Comma-separated ad IDs")
-@click.option("--campaign-ids", help="Comma-separated campaign IDs")
-@click.option("--adgroup-ids", help="Comma-separated ad group IDs")
-@click.option("--status", help="Filter by status")
-@click.option("--statuses", help="Comma-separated statuses")
-@click.option("--states", help="Comma-separated states")
-@click.option("--types", help="Comma-separated ad types")
-@click.option("--mobile", type=click.Choice(["YES", "NO"], case_sensitive=False))
-@click.option("--vcard-ids", help="Comma-separated vCard IDs")
-@click.option("--sitelink-set-ids", help="Comma-separated sitelink set IDs")
-@click.option("--image-hashes", help="Comma-separated ad image hashes")
-@click.option(
-    "--vcard-moderation-statuses", help="Comma-separated vCard moderation statuses"
-)
-@click.option(
-    "--sitelinks-moderation-statuses",
-    help="Comma-separated sitelinks moderation statuses",
-)
-@click.option(
-    "--image-moderation-statuses", help="Comma-separated image moderation statuses"
-)
-@click.option("--adextension-ids", help="Comma-separated ad extension IDs")
-@click.option("--limit", type=int, help="Limit number of results")
-@click.option("--fetch-all", is_flag=True, help="Fetch all pages")
-@click.option("--format", "output_format", default="json", help="Output format")
-@click.option("--output", help="Output file")
-@click.option("--fields", help="Comma-separated top-level field names")
-@click.option(
-    "--cpc-video-ad-builder-ad-field-names",
-    help=(
-        "Comma-separated CpcVideoAdBuilderAdFieldNames "
-        "(e.g. CreativeId,Href). Sent as separate top-level request "
-        "parameter per the AdsGetRequest WSDL."
-    ),
-)
-@click.option(
-    "--cpm-banner-ad-builder-ad-field-names",
-    help=(
-        "Comma-separated CpmBannerAdBuilderAdFieldNames "
-        "(e.g. CreativeId,Href). Sent as separate top-level request "
-        "parameter per the AdsGetRequest WSDL."
-    ),
-)
-@click.option(
-    "--cpm-video-ad-builder-ad-field-names",
-    help=(
-        "Comma-separated CpmVideoAdBuilderAdFieldNames "
-        "(e.g. CreativeId,Href). Sent as separate top-level request "
-        "parameter per the AdsGetRequest WSDL."
-    ),
-)
-@click.option(
-    "--dynamic-text-ad-field-names",
-    help=(
-        "Comma-separated DynamicTextAdFieldNames (e.g. Title,Text,Href). "
-        "Sent as separate top-level request parameter per the "
-        "AdsGetRequest WSDL."
-    ),
-)
-@click.option(
-    "--listing-ad-field-names",
-    help=(
-        "Comma-separated ListingAdFieldNames (e.g. Title,Text,Href). "
-        "Sent as separate top-level request parameter per the "
-        "AdsGetRequest WSDL."
-    ),
-)
-@click.option(
-    "--mobile-app-ad-builder-ad-field-names",
-    help=(
-        "Comma-separated MobileAppAdBuilderAdFieldNames "
-        "(e.g. CreativeId,TrackingUrl). Sent as separate top-level "
-        "request parameter per the AdsGetRequest WSDL."
-    ),
-)
-@click.option(
-    "--mobile-app-ad-field-names",
-    help=(
-        "Comma-separated MobileAppAdFieldNames (e.g. Title,Text,TrackingUrl). "
-        "Sent as separate top-level request parameter per the "
-        "AdsGetRequest WSDL."
-    ),
-)
-@click.option(
-    "--mobile-app-cpc-video-ad-builder-ad-field-names",
-    help=(
-        "Comma-separated MobileAppCpcVideoAdBuilderAdFieldNames "
-        "(e.g. CreativeId,TrackingUrl). Sent as separate top-level "
-        "request parameter per the AdsGetRequest WSDL."
-    ),
-)
-@click.option(
-    "--mobile-app-image-ad-field-names",
-    help=(
-        "Comma-separated MobileAppImageAdFieldNames (e.g. ImageHash,TrackingUrl). "
-        "Sent as separate top-level request parameter per the "
-        "AdsGetRequest WSDL."
-    ),
-)
-@click.option(
-    "--responsive-ad-field-names",
-    help=(
-        "Comma-separated ResponsiveAdFieldNames (e.g. Titles,Texts,Href). "
-        "Sent as separate top-level request parameter per the "
-        "AdsGetRequest WSDL."
-    ),
-)
-@click.option(
-    "--shopping-ad-field-names",
-    help=(
-        "Comma-separated ShoppingAdFieldNames (e.g. Titles,Texts,Href). "
-        "Sent as separate top-level request parameter per the "
-        "AdsGetRequest WSDL."
-    ),
-)
-@click.option(
-    "--smart-ad-builder-ad-field-names",
-    help=(
-        "Comma-separated SmartAdBuilderAdFieldNames (e.g. CreativeId). "
-        "Sent as separate top-level request parameter per the "
-        "AdsGetRequest WSDL."
-    ),
-)
-@click.option(
-    "--text-ad-builder-ad-field-names",
-    help=(
-        "Comma-separated TextAdBuilderAdFieldNames (e.g. CreativeId,Href). "
-        "Sent as separate top-level request parameter per the "
-        "AdsGetRequest WSDL."
-    ),
-)
-@click.option(
-    "--text-ad-field-names",
-    help=(
-        "Comma-separated TextAdFieldNames (e.g. Title,Text,Href). "
-        "Sent as separate top-level request parameter per the "
-        "AdsGetRequest WSDL."
-    ),
-)
-@click.option(
-    "--text-ad-price-extension-field-names",
-    help=(
-        "Comma-separated TextAdPriceExtensionFieldNames "
-        "(e.g. Price,OldPrice,PriceQualifier). Sent as separate top-level "
-        "request parameter per the AdsGetRequest WSDL."
-    ),
-)
-@click.option(
-    "--text-image-ad-field-names",
-    help=(
-        "Comma-separated TextImageAdFieldNames (e.g. Href,ImageHash). "
-        "Sent as separate top-level request parameter per the "
-        "AdsGetRequest WSDL."
-    ),
-)
-@click.option("--dry-run", is_flag=True, help="Show request without sending")
-@click.pass_context
-@handle_api_errors
-def get(
-    ctx,
-    ids,
-    campaign_ids,
-    adgroup_ids,
-    status,
-    statuses,
-    states,
-    types,
-    mobile,
-    vcard_ids,
-    sitelink_set_ids,
-    image_hashes,
-    vcard_moderation_statuses,
-    sitelinks_moderation_statuses,
-    image_moderation_statuses,
-    adextension_ids,
-    limit,
-    fetch_all,
-    output_format,
-    output,
-    fields,
-    cpc_video_ad_builder_ad_field_names,
-    cpm_banner_ad_builder_ad_field_names,
-    cpm_video_ad_builder_ad_field_names,
-    dynamic_text_ad_field_names,
-    listing_ad_field_names,
-    mobile_app_ad_builder_ad_field_names,
-    mobile_app_ad_field_names,
-    mobile_app_cpc_video_ad_builder_ad_field_names,
-    mobile_app_image_ad_field_names,
-    responsive_ad_field_names,
-    shopping_ad_field_names,
-    smart_ad_builder_ad_field_names,
-    text_ad_builder_ad_field_names,
-    text_ad_field_names,
-    text_ad_price_extension_field_names,
-    text_image_ad_field_names,
-    dry_run,
-):
-    """Get ads"""
-    if status and statuses:
-        raise click.UsageError(t("--status and --statuses are mutually exclusive"))
-
-    field_names = parse_csv_strings(fields) or get_default_fields("ads", "FieldNames")
-
-    raw_nested = (
-        (
-            "CpcVideoAdBuilderAdFieldNames",
-            cpc_video_ad_builder_ad_field_names,
-        ),
-        (
-            "CpmBannerAdBuilderAdFieldNames",
-            cpm_banner_ad_builder_ad_field_names,
-        ),
-        (
-            "CpmVideoAdBuilderAdFieldNames",
-            cpm_video_ad_builder_ad_field_names,
-        ),
-        ("DynamicTextAdFieldNames", dynamic_text_ad_field_names),
-        ("ListingAdFieldNames", listing_ad_field_names),
-        ("MobileAppAdBuilderAdFieldNames", mobile_app_ad_builder_ad_field_names),
-        ("MobileAppAdFieldNames", mobile_app_ad_field_names),
-        (
-            "MobileAppCpcVideoAdBuilderAdFieldNames",
-            mobile_app_cpc_video_ad_builder_ad_field_names,
-        ),
-        ("MobileAppImageAdFieldNames", mobile_app_image_ad_field_names),
-        ("ResponsiveAdFieldNames", responsive_ad_field_names),
-        ("ShoppingAdFieldNames", shopping_ad_field_names),
-        ("SmartAdBuilderAdFieldNames", smart_ad_builder_ad_field_names),
-        ("TextAdBuilderAdFieldNames", text_ad_builder_ad_field_names),
-        ("TextAdFieldNames", text_ad_field_names),
-        (
-            "TextAdPriceExtensionFieldNames",
-            text_ad_price_extension_field_names,
-        ),
-        ("TextImageAdFieldNames", text_image_ad_field_names),
-    )
-    parsed_nested = parse_nested_field_names(raw_nested)
-    parsed_nested.setdefault(
-        "TextAdFieldNames", get_default_fields("ads", "TextAdFieldNames")
-    )
-
-    criteria: dict[str, Any] = {}
-    if ids:
-        criteria["Ids"] = parse_ids(ids)
-    if campaign_ids:
-        criteria["CampaignIds"] = parse_ids(campaign_ids)
-    if adgroup_ids:
-        criteria["AdGroupIds"] = parse_ids(adgroup_ids)
-    if status:
-        criteria["Statuses"] = [status]
-    add_criteria_csv(criteria, "Statuses", statuses, upper=True)
-    add_criteria_csv(criteria, "States", states, upper=True)
-    add_criteria_csv(criteria, "Types", types, upper=True)
-    if mobile:
-        criteria["Mobile"] = mobile.upper()
-    add_criteria_csv(criteria, "VCardIds", vcard_ids, integers=True)
-    add_criteria_csv(criteria, "SitelinkSetIds", sitelink_set_ids, integers=True)
-    add_criteria_csv(criteria, "AdImageHashes", image_hashes)
-    add_criteria_csv(
-        criteria, "VCardModerationStatuses", vcard_moderation_statuses, upper=True
-    )
-    add_criteria_csv(
-        criteria,
-        "SitelinksModerationStatuses",
-        sitelinks_moderation_statuses,
-        upper=True,
-    )
-    add_criteria_csv(
-        criteria, "AdImageModerationStatuses", image_moderation_statuses, upper=True
-    )
-    add_criteria_csv(criteria, "AdExtensionIds", adextension_ids, integers=True)
-
-    enforce_criteria_array_limits(
-        criteria, ADS_GET_CRITERIA_LIMITS, command_name="ads get"
-    )
-
-    if not criteria:
-        raise click.UsageError(t("Provide at least one typed filter"))
-
-    params = build_common_params(
-        criteria=criteria, field_names=field_names, limit=limit
-    )
-    params.update(parsed_nested)
-
-    body = {"method": "get", "params": params}
-
-    if dry_run:
-        format_output(body, "json", None)
-        return
-
-    client = client_from_ctx(ctx, create_client)
-
-    result = client.ads().post(data=body)
-
-    if fetch_all:
-        items = []
-        for item in result().iter_items():
-            items.append(item)
-        format_output(items, output_format, output)
-    else:
-        data = result().extract()
-        format_output(data, output_format, output)
-
-
-# dest-name -> CLI flag label, used by build_ad_object's incompatible-flag
-# guard and by the batch row normalizer (issue #562). Module-level so both the
-# single-flag command and the --from-file path share one source of truth.
-_ADS_ADD_FLAG_FOR = {
-    "title": "--title",
-    "text": "--text",
-    "titles": "--titles",
-    "texts": "--texts",
-    "href": "--href",
-    "image_hash": "--image-hash",
-    "image_hashes": "--image-hashes",
-    "action": "--action",
-    "tracking_url": "--tracking-url",
-    "age_label": "--age-label",
-    "mobile_app_features": "--mobile-app-feature",
-    "title2": "--title2",
-    "display_url_path": "--display-url-path",
-    "mobile": "--mobile",
-    "vcard_id": "--vcard-id",
-    "sitelink_set_id": "--sitelink-set-id",
-    "turbo_page_id": "--turbo-page-id",
-    "ad_extensions": "--ad-extensions",
-    "final_url": "--final-url",
-    "video_extension_creative_id": "--video-extension-creative-id",
-    "price_extension_price": "--price-extension-price",
-    "price_extension_old_price": "--price-extension-old-price",
-    "price_extension_price_qualifier": "--price-extension-price-qualifier",
-    "price_extension_price_currency": "--price-extension-price-currency",
-    "video_extension_ids": "--video-extension-ids",
-    "business_id": "--business-id",
-    "prefer_vcard_over_business": "--prefer-vcard-over-business",
-    "erir_ad_description": "--erir-ad-description",
-    "creative_id": "--creative-id",
-    "tracking_pixels": "--tracking-pixels",
-    "logo_extension_hash": "--logo-extension-hash",
-    "feed_id": "--feed-id",
-    "feed_filter_conditions": "--feed-filter-condition",
-    "title_sources": "--title-sources",
-    "text_sources": "--text-sources",
-    "default_texts": "--default-texts",
-}
-
-
-# Static lookup tables for build_ad_object — hoisted to module level so they are
-# allocated once at import, not rebuilt per ad (it runs once per row in batch).
-_ADS_ADD_SUPPORTED_TYPES = frozenset(
-    {
-        "TEXT_AD",
-        "TEXT_IMAGE_AD",
-        "MOBILE_APP_AD",
-        "DYNAMIC_TEXT_AD",
-        "MOBILE_APP_IMAGE_AD",
-        "RESPONSIVE_AD",
-        "SHOPPING_AD",
-        "LISTING_AD",
-        "SMART_AD_BUILDER_AD",
-        *AD_BUILDER_ADD_BLOCKS,
-    }
-)
-_ADS_ADD_TYPE_FIELDS = {
-    "TEXT_AD": {
-        "title",
-        "text",
-        "href",
-        "image_hash",
-        "title2",
-        "display_url_path",
-        "mobile",
-        "vcard_id",
-        "sitelink_set_id",
-        "turbo_page_id",
-        "ad_extensions",
-        "final_url",
-        "video_extension_creative_id",
-        "price_extension_price",
-        "price_extension_old_price",
-        "price_extension_price_qualifier",
-        "price_extension_price_currency",
-        "business_id",
-        "prefer_vcard_over_business",
-        "erir_ad_description",
-    },
-    "TEXT_IMAGE_AD": {
-        "href",
-        "image_hash",
-        "turbo_page_id",
-        "final_url",
-        "erir_ad_description",
-    },
-    "RESPONSIVE_AD": {
-        "texts",
-        "titles",
-        "href",
-        "age_label",
-        "display_url_path",
-        "image_hashes",
-        "sitelink_set_id",
-        "ad_extensions",
-        "video_extension_ids",
-        "price_extension_price",
-        "price_extension_old_price",
-        "price_extension_price_qualifier",
-        "price_extension_price_currency",
-        "business_id",
-        "erir_ad_description",
-    },
-    "SHOPPING_AD": FEED_BASED_ADD_FIELDS,
-    "LISTING_AD": FEED_BASED_ADD_FIELDS,
-    **AD_BUILDER_ADD_TYPE_FIELDS,
-    "DYNAMIC_TEXT_AD": DYNAMIC_TEXT_AD_ADD_FIELDS,
-    "MOBILE_APP_AD": MOBILE_APP_AD_ADD_FIELDS,
-    "MOBILE_APP_IMAGE_AD": MOBILE_APP_IMAGE_AD_ADD_FIELDS,
-    "SMART_AD_BUILDER_AD": SMART_AD_BUILDER_ADD_FIELDS,
-}
-
-
-def build_ad_object(*, adgroup_id, ad_type, mobile_provided, flags, flag_for=None):
-    """Build a single ``Ads`` item dict from flag values (issue #562).
-
-    Pure (no ``ctx``, no I/O): performs ``--type`` validation, the
-    incompatible-flag guard, per-subtype dispatch, and per-subtype required-field
-    checks, returning ``{"AdGroupId": ..., "<SubType>": {...}}``. Both the
-    single-flag ``ads add`` command and the ``--from-file`` batch normalizer call
-    it so they emit byte-identical objects.
-
-    ``flags`` is keyed by the command's dest var names (``image_hash``,
-    ``mobile_app_features``, ...); missing keys default to ``None``.
-    ``mobile_provided`` is the explicitly-passed ``--mobile`` value (``None`` when
-    the flag was not given), already resolved by the caller (it needs ``ctx``).
-    """
-    flag_for = flag_for if flag_for is not None else _ADS_ADD_FLAG_FOR
-
-    # Unpack flags into locals so the dispatch body below is byte-identical to
-    # the historical inline command body.
-    title = flags.get("title")
-    text = flags.get("text")
-    titles = flags.get("titles")
-    texts = flags.get("texts")
-    href = flags.get("href")
-    image_hash = flags.get("image_hash")
-    image_hashes = flags.get("image_hashes")
-    action = flags.get("action")
-    tracking_url = flags.get("tracking_url")
-    age_label = flags.get("age_label")
-    mobile_app_features = flags.get("mobile_app_features") or ()
-    title2 = flags.get("title2")
-    display_url_path = flags.get("display_url_path")
-    vcard_id = flags.get("vcard_id")
-    sitelink_set_id = flags.get("sitelink_set_id")
-    turbo_page_id = flags.get("turbo_page_id")
-    ad_extensions = flags.get("ad_extensions")
-    final_url = flags.get("final_url")
-    video_extension_creative_id = flags.get("video_extension_creative_id")
-    price_extension_price = flags.get("price_extension_price")
-    price_extension_old_price = flags.get("price_extension_old_price")
-    price_extension_price_qualifier = flags.get("price_extension_price_qualifier")
-    price_extension_price_currency = flags.get("price_extension_price_currency")
-    video_extension_ids = flags.get("video_extension_ids")
-    business_id = flags.get("business_id")
-    prefer_vcard_over_business = flags.get("prefer_vcard_over_business")
-    erir_ad_description = flags.get("erir_ad_description")
-    creative_id = flags.get("creative_id")
-    tracking_pixels = flags.get("tracking_pixels")
-    logo_extension_hash = flags.get("logo_extension_hash")
-    feed_id = flags.get("feed_id")
-    feed_filter_conditions = flags.get("feed_filter_conditions") or ()
-    title_sources = flags.get("title_sources")
-    text_sources = flags.get("text_sources")
-    default_texts = flags.get("default_texts")
-    # --mobile defaults to "NO" in the payload; the guard only sees it when
-    # explicitly provided (mobile_provided).
-    mobile = mobile_provided or "NO"
-
-    ad_type_norm = (ad_type or "TEXT_AD").upper().replace("-", "_")
-    if ad_type_norm not in _ADS_ADD_SUPPORTED_TYPES:
-        raise click.UsageError(
-            t(
-                "Invalid value for '--type': {ad_type!r} is not one of 'TEXT_AD', 'TEXT_IMAGE_AD', 'MOBILE_APP_AD', 'DYNAMIC_TEXT_AD', 'MOBILE_APP_IMAGE_AD', 'RESPONSIVE_AD', 'SHOPPING_AD', 'LISTING_AD', 'SMART_AD_BUILDER_AD', 'TEXT_AD_BUILDER_AD', 'MOBILE_APP_AD_BUILDER_AD', 'MOBILE_APP_CPC_VIDEO_AD_BUILDER_AD', 'CPC_VIDEO_AD_BUILDER_AD', 'CPM_BANNER_AD_BUILDER_AD', 'CPM_VIDEO_AD_BUILDER_AD'."
-            ).format(ad_type=ad_type)
-        )
-
-    # `provided` is `flags` with the explicit --mobile value (the command passes
-    # mobile separately; a batch row pops it out). _reject_incompatible_flags
-    # skips None/() values and absent keys alike, so a sparse dict is fine.
-    provided = {**flags, "mobile": mobile_provided}
-    _reject_incompatible_flags(
-        ad_type_norm, _ADS_ADD_TYPE_FIELDS[ad_type_norm], provided, flag_for
-    )
-
-    ad_data = {"AdGroupId": adgroup_id}
-    if ad_type_norm == "TEXT_AD":
-        missing_fields = [
-            option_name
-            for option_name, value in (
-                ("--title", title),
-                ("--text", text),
-                ("--href", href),
-            )
-            if not value
-        ]
-        if missing_fields:
-            raise click.UsageError(
-                t("TEXT_AD requires {arg0}").format(arg0=", ".join(missing_fields))
-            )
-        text_ad = {
-            "Mobile": mobile.upper(),
-            "Title": title,
-            "Text": text,
-            "Href": href,
-        }
-        if image_hash:
-            text_ad["AdImageHash"] = image_hash
-        if title2:
-            text_ad["Title2"] = title2
-        if display_url_path:
-            text_ad["DisplayUrlPath"] = display_url_path
-        if vcard_id:
-            text_ad["VCardId"] = vcard_id
-        if sitelink_set_id:
-            text_ad["SitelinkSetId"] = sitelink_set_id
-        if turbo_page_id:
-            text_ad["TurboPageId"] = turbo_page_id
-        if ad_extensions:
-            text_ad["AdExtensionIds"] = parse_ids(ad_extensions)
-        if final_url:
-            text_ad["FinalUrl"] = final_url
-        if video_extension_creative_id is not None:
-            text_ad["VideoExtension"] = {"CreativeId": video_extension_creative_id}
-        price_extension = _build_price_extension_add(
-            price_extension_price,
-            price_extension_old_price,
-            price_extension_price_qualifier,
-            price_extension_price_currency,
-        )
-        if price_extension:
-            text_ad["PriceExtension"] = price_extension
-        if business_id is not None:
-            text_ad["BusinessId"] = business_id
-        if prefer_vcard_over_business:
-            text_ad["PreferVCardOverBusiness"] = prefer_vcard_over_business.upper()
-        if erir_ad_description:
-            text_ad["ErirAdDescription"] = erir_ad_description
-        ad_data["TextAd"] = text_ad
-    elif ad_type_norm == "DYNAMIC_TEXT_AD":
-        ad_data["DynamicTextAd"] = _build_dynamic_text_ad_add(
-            text,
-            image_hash,
-            vcard_id,
-            sitelink_set_id,
-            ad_extensions,
-        )
-    elif ad_type_norm == "TEXT_IMAGE_AD":
-        if title or text:
-            raise click.UsageError(
-                t(
-                    "--title/--text are only valid for TEXT_AD. "
-                    "For TEXT_IMAGE_AD, use --image-hash and "
-                    "--href / --turbo-page-id."
-                )
-            )
-        if not image_hash:
-            raise click.UsageError(t("TEXT_IMAGE_AD requires --image-hash"))
-        if not href and turbo_page_id is None:
-            raise click.UsageError(
-                t("TEXT_IMAGE_AD requires either --href or --turbo-page-id.")
-            )
-        text_image_ad = {"AdImageHash": image_hash}
-        if erir_ad_description:
-            text_image_ad["ErirAdDescription"] = erir_ad_description
-        if final_url:
-            text_image_ad["FinalUrl"] = final_url
-        if href:
-            text_image_ad["Href"] = href
-        if turbo_page_id is not None:
-            text_image_ad["TurboPageId"] = turbo_page_id
-        ad_data["TextImageAd"] = text_image_ad
-    elif ad_type_norm == "RESPONSIVE_AD":
-        missing_fields = [
-            option_name
-            for option_name, value in (
-                ("--texts", texts),
-                ("--titles", titles),
-            )
-            if value is None
-        ]
-        if missing_fields:
-            raise click.UsageError(
-                t("RESPONSIVE_AD requires {arg0}").format(
-                    arg0=", ".join(missing_fields)
-                )
-            )
-        if not href and business_id is None:
-            raise click.UsageError(
-                t("RESPONSIVE_AD requires either --href or --business-id.")
-            )
-
-        parsed_texts = _parse_required_csv_strings(texts, "--texts")
-        parsed_titles = _parse_required_csv_strings(titles, "--titles")
-        responsive_ad: dict[str, Any] = {
-            "Texts": parsed_texts,
-            "Titles": parsed_titles,
-        }
-        parsed_image_hashes = _parse_required_csv_strings(
-            image_hashes, "--image-hashes"
-        )
-        if parsed_image_hashes:
-            responsive_ad["AdImageHashes"] = parsed_image_hashes
-        parsed_video_extension_ids = _parse_required_ids(
-            video_extension_ids, "--video-extension-ids"
-        )
-        if parsed_video_extension_ids:
-            responsive_ad["VideoExtensionIds"] = parsed_video_extension_ids
-        if sitelink_set_id is not None:
-            responsive_ad["SitelinkSetId"] = sitelink_set_id
-        if ad_extensions:
-            responsive_ad["AdExtensionIds"] = parse_ids(ad_extensions)
-        if href:
-            responsive_ad["Href"] = href
-        if age_label:
-            responsive_ad["AgeLabel"] = age_label.upper()
-        if display_url_path:
-            responsive_ad["DisplayUrlPath"] = display_url_path
-        price_extension = _build_price_extension_add(
-            price_extension_price,
-            price_extension_old_price,
-            price_extension_price_qualifier,
-            price_extension_price_currency,
-            container_name="ResponsiveAd",
-        )
-        if price_extension:
-            responsive_ad["PriceExtension"] = price_extension
-        if business_id is not None:
-            responsive_ad["BusinessId"] = business_id
-        if erir_ad_description:
-            responsive_ad["ErirAdDescription"] = erir_ad_description
-        ad_data["ResponsiveAd"] = responsive_ad
-    elif ad_type_norm in {"SHOPPING_AD", "LISTING_AD"}:
-        field_name = "ShoppingAd" if ad_type_norm == "SHOPPING_AD" else "ListingAd"
-        ad_data[field_name] = _build_feed_based_ad_add(
-            feed_id,
-            default_texts,
-            sitelink_set_id,
-            ad_extensions,
-            business_id,
-            feed_filter_conditions,
-            title_sources,
-            text_sources,
-            field_name,
-        )
-    elif ad_type_norm in AD_BUILDER_ADD_BLOCKS:
-        field_name = AD_BUILDER_ADD_BLOCKS[ad_type_norm]
-        ad_data[field_name] = _build_ad_builder_add(
-            creative_id,
-            erir_ad_description,
-            final_url,
-            href,
-            turbo_page_id,
-            tracking_url,
-            tracking_pixels,
-            ad_type_norm,
-            field_name,
-        )
-    elif ad_type_norm == "MOBILE_APP_AD":
-        if href:
-            raise click.UsageError(
-                t(
-                    "--href does not apply to MOBILE_APP_AD. "
-                    "Use --tracking-url instead."
-                )
-            )
-        missing_fields = [
-            option_name
-            for option_name, value in (
-                ("--title", title),
-                ("--text", text),
-                ("--action", action),
-            )
-            if not value
-        ]
-        if missing_fields:
-            raise click.UsageError(
-                t("MOBILE_APP_AD requires {arg0}").format(
-                    arg0=", ".join(missing_fields)
-                )
-            )
-        mobile_app_ad = {
-            "Title": title,
-            "Text": text,
-            "Action": action.upper(),
-        }
-        if image_hash:
-            mobile_app_ad["AdImageHash"] = image_hash
-        if tracking_url:
-            mobile_app_ad["TrackingUrl"] = tracking_url
-        if age_label:
-            mobile_app_ad["AgeLabel"] = age_label.upper()
-        parsed_features = _parse_mobile_app_features(mobile_app_features)
-        if parsed_features:
-            mobile_app_ad["Features"] = parsed_features
-        if video_extension_creative_id is not None:
-            mobile_app_ad["VideoExtension"] = {
-                "CreativeId": video_extension_creative_id
-            }
-        if erir_ad_description:
-            mobile_app_ad["ErirAdDescription"] = erir_ad_description
-        ad_data["MobileAppAd"] = mobile_app_ad
-    elif ad_type_norm == "MOBILE_APP_IMAGE_AD":
-        ad_data["MobileAppImageAd"] = _build_mobile_app_image_ad_add(
-            image_hash,
-            erir_ad_description,
-            tracking_url,
-        )
-    elif ad_type_norm == "SMART_AD_BUILDER_AD":
-        ad_data["SmartAdBuilderAd"] = _build_smart_ad_builder_ad_add(
-            logo_extension_hash,
-        )
-
-    return ad_data
-
-
-# dest -> "--flag" map for the `ads update` flag set (mirrors _ADS_ADD_FLAG_FOR).
-# Hoisted to module level so build_ad_update_object and the batch normalizer
-# share one source of truth.
-_ADS_UPDATE_FLAG_FOR = {
-    "title": "--title",
-    "text": "--text",
-    "titles": "--titles",
-    "texts": "--texts",
-    "href": "--href",
-    "image_hash": "--image-hash",
-    "clear_image_hash": "--clear-image-hash",
-    "image_hashes": "--image-hashes",
-    "action": "--action",
-    "tracking_url": "--tracking-url",
-    "age_label": "--age-label",
-    "mobile_app_features": "--mobile-app-feature",
-    "title2": "--title2",
-    "display_url_path": "--display-url-path",
-    "vcard_id": "--vcard-id",
-    "sitelink_set_id": "--sitelink-set-id",
-    "turbo_page_id": "--turbo-page-id",
-    "callouts_add": "--callouts-add",
-    "callouts_remove": "--callouts-remove",
-    "callouts_set": "--callouts-set",
-    "video_extension_creative_id": "--video-extension-creative-id",
-    "video_extension_ids": "--video-extension-ids",
-    "price_extension_price": "--price-extension-price",
-    "price_extension_old_price": "--price-extension-old-price",
-    "price_extension_price_qualifier": "--price-extension-price-qualifier",
-    "price_extension_price_currency": "--price-extension-price-currency",
-    "business_id": "--business-id",
-    "prefer_vcard_over_business": "--prefer-vcard-over-business",
-    "erir_ad_description": "--erir-ad-description",
-    "logo_extension_hash": "--logo-extension-hash",
-    "creative_id": "--creative-id",
-    "creative_erir_ad_description": "--creative-erir-ad-description",
-    "final_url": "--final-url",
-    "tracking_pixels": "--tracking-pixels",
-    "feed_filter_conditions": "--feed-filter-condition",
-    "title_sources": "--title-sources",
-    "text_sources": "--text-sources",
-    "default_texts": "--default-texts",
-}
-
-
-def _ads_update_type_fields():
-    """Per-WSDL-subtype field allow-list for ``ads update``.
-
-    Built once at import (not per ad) so the batch path, which runs this once per
-    row, does not rebuild the dict. Each ``--type`` accepts only the options that
-    map to fields inside its ``AdUpdateItem`` subtype; a flag outside the
-    allow-list would be silently dropped, so the caller rejects it up front
-    (issue #198 H2).
-    """
-    return {
-        "TEXT_AD": TEXT_AD_UPDATE_FIELDS,
-        "DYNAMIC_TEXT_AD": {
-            "text",
-            "image_hash",
-            "clear_image_hash",
-            "vcard_id",
-            "sitelink_set_id",
-            "callouts_add",
-            "callouts_remove",
-            "callouts_set",
-        },
-        "TEXT_IMAGE_AD": TEXT_IMAGE_AD_UPDATE_FIELDS,
-        "MOBILE_APP_AD": MOBILE_APP_AD_UPDATE_FIELDS,
-        "MOBILE_APP_IMAGE_AD": MOBILE_APP_IMAGE_UPDATE_FIELDS,
-        "RESPONSIVE_AD": {
-            "texts",
-            "titles",
-            "image_hashes",
-            "video_extension_ids",
-            "sitelink_set_id",
-            "callouts_add",
-            "callouts_remove",
-            "callouts_set",
-            "href",
-            "age_label",
-            "display_url_path",
-            "price_extension_price",
-            "price_extension_old_price",
-            "price_extension_price_qualifier",
-            "price_extension_price_currency",
-            "business_id",
-            "erir_ad_description",
-        },
-        "SHOPPING_AD": FEED_BASED_UPDATE_FIELDS,
-        "LISTING_AD": FEED_BASED_UPDATE_FIELDS,
-        "SMART_AD_BUILDER_AD": SMART_AD_BUILDER_UPDATE_FIELDS,
-        **AD_BUILDER_TYPE_FIELDS,
-    }
-
-
-_ADS_UPDATE_TYPE_FIELDS = _ads_update_type_fields()
-_ADS_UPDATE_SUPPORTED_TYPES = frozenset(
-    {
-        "TEXT_AD",
-        "TEXT_IMAGE_AD",
-        "MOBILE_APP_AD",
-        "MOBILE_APP_IMAGE_AD",
-        "DYNAMIC_TEXT_AD",
-        "RESPONSIVE_AD",
-        "SHOPPING_AD",
-        "LISTING_AD",
-        "SMART_AD_BUILDER_AD",
-        *AD_BUILDER_UPDATE_BLOCKS,
-    }
-)
-
-
-def build_ad_update_object(*, ad_id, ad_type, flags, flag_for=None):
-    """Build a single ``Ads`` update item dict from flag values (issue #563).
-
-    Pure (no ``ctx``, no I/O): performs ``--type`` validation, the
-    incompatible-flag / "does not convert between subtypes" guard, per-subtype
-    assembly, and the empty-subtype no-op guard, returning
-    ``{"Id": ..., "<SubType>": {...}}``. Both the single-flag ``ads update``
-    command and the ``--from-file`` batch normalizer call it so they emit
-    byte-identical objects.
-
-    ``flags`` is keyed by the command's dest var names (``image_hash``,
-    ``clear_image_hash``, ``mobile_app_features``, ...); missing keys default to
-    ``None``. The caller is responsible for the ``--status`` rejection and the
-    ``--image-hash``/``--clear-image-hash`` mutex (those have command-level
-    wording).
-    """
-    flag_for = flag_for if flag_for is not None else _ADS_UPDATE_FLAG_FOR
-
-    # Unpack flags into locals so the dispatch body below is byte-identical to
-    # the historical inline command body.
-    title = flags.get("title")
-    text = flags.get("text")
-    titles = flags.get("titles")
-    texts = flags.get("texts")
-    href = flags.get("href")
-    image_hash = flags.get("image_hash")
-    clear_image_hash = flags.get("clear_image_hash")
-    image_hashes = flags.get("image_hashes")
-    action = flags.get("action")
-    tracking_url = flags.get("tracking_url")
-    age_label = flags.get("age_label")
-    mobile_app_features = flags.get("mobile_app_features") or ()
-    title2 = flags.get("title2")
-    display_url_path = flags.get("display_url_path")
-    vcard_id = flags.get("vcard_id")
-    sitelink_set_id = flags.get("sitelink_set_id")
-    turbo_page_id = flags.get("turbo_page_id")
-    callouts_add = flags.get("callouts_add")
-    callouts_remove = flags.get("callouts_remove")
-    callouts_set = flags.get("callouts_set")
-    video_extension_creative_id = flags.get("video_extension_creative_id")
-    video_extension_ids = flags.get("video_extension_ids")
-    price_extension_price = flags.get("price_extension_price")
-    price_extension_old_price = flags.get("price_extension_old_price")
-    price_extension_price_qualifier = flags.get("price_extension_price_qualifier")
-    price_extension_price_currency = flags.get("price_extension_price_currency")
-    business_id = flags.get("business_id")
-    prefer_vcard_over_business = flags.get("prefer_vcard_over_business")
-    erir_ad_description = flags.get("erir_ad_description")
-    logo_extension_hash = flags.get("logo_extension_hash")
-    creative_id = flags.get("creative_id")
-    creative_erir_ad_description = flags.get("creative_erir_ad_description")
-    final_url = flags.get("final_url")
-    tracking_pixels = flags.get("tracking_pixels")
-    feed_filter_conditions = flags.get("feed_filter_conditions") or ()
-    title_sources = flags.get("title_sources")
-    text_sources = flags.get("text_sources")
-    default_texts = flags.get("default_texts")
-
-    ad_type_norm = ad_type.upper().replace("-", "_")
-    if ad_type_norm not in _ADS_UPDATE_SUPPORTED_TYPES:
-        raise click.UsageError(
-            t(
-                "Invalid value for '--type': {ad_type!r} is not one of 'TEXT_AD', 'TEXT_IMAGE_AD', 'MOBILE_APP_AD', 'DYNAMIC_TEXT_AD', 'MOBILE_APP_IMAGE_AD', 'RESPONSIVE_AD', 'SHOPPING_AD', 'LISTING_AD', 'SMART_AD_BUILDER_AD', 'TEXT_AD_BUILDER_AD', 'MOBILE_APP_AD_BUILDER_AD', 'MOBILE_APP_CPC_VIDEO_AD_BUILDER_AD', 'CPC_VIDEO_AD_BUILDER_AD', 'CPM_BANNER_AD_BUILDER_AD', 'CPM_VIDEO_AD_BUILDER_AD'."
-            ).format(ad_type=ad_type)
-        )
-
-    provided = {
-        "title": title,
-        "text": text,
-        "titles": titles,
-        "texts": texts,
-        "href": href,
-        "image_hash": image_hash,
-        "clear_image_hash": clear_image_hash or None,
-        "image_hashes": image_hashes,
-        "action": action,
-        "tracking_url": tracking_url,
-        "age_label": age_label,
-        "mobile_app_features": mobile_app_features,
-        "title2": title2,
-        "display_url_path": display_url_path,
-        "vcard_id": vcard_id,
-        "sitelink_set_id": sitelink_set_id,
-        "turbo_page_id": turbo_page_id,
-        "callouts_add": callouts_add,
-        "callouts_remove": callouts_remove,
-        "callouts_set": callouts_set,
-        "video_extension_creative_id": video_extension_creative_id,
-        "video_extension_ids": video_extension_ids,
-        "price_extension_price": price_extension_price,
-        "price_extension_old_price": price_extension_old_price,
-        "price_extension_price_qualifier": price_extension_price_qualifier,
-        "price_extension_price_currency": price_extension_price_currency,
-        "business_id": business_id,
-        "prefer_vcard_over_business": prefer_vcard_over_business,
-        "erir_ad_description": erir_ad_description,
-        "logo_extension_hash": logo_extension_hash,
-        "creative_id": creative_id,
-        "creative_erir_ad_description": creative_erir_ad_description,
-        "final_url": final_url,
-        "tracking_pixels": tracking_pixels,
-        "feed_filter_conditions": feed_filter_conditions,
-        "title_sources": title_sources,
-        "text_sources": text_sources,
-        "default_texts": default_texts,
-    }
-    try:
-        _reject_incompatible_flags(
-            ad_type_norm, _ADS_UPDATE_TYPE_FIELDS[ad_type_norm], provided, flag_for
-        )
-    except click.UsageError as exc:
-        raise click.UsageError(
-            t(
-                "{arg0} --type selects the existing ad subtype update block; it does not convert an ad between subtypes."
-            ).format(arg0=exc.message)
-        )
-
-    # Validate up-front so SET vs ADD/REMOVE mutex errors raise UsageError
-    # before any payload work, bypassing the generic ``except Exception``
-    # net wrapped around the network call below.
-    callout_setting = _build_callout_setting(
-        callouts_add, callouts_remove, callouts_set
-    )
-    price_extension = _build_price_extension(
-        price_extension_price,
-        price_extension_old_price,
-        price_extension_price_qualifier,
-        price_extension_price_currency,
-    )
-    parsed_mobile_app_features = _parse_mobile_app_features(mobile_app_features)
-
-    ad_data = {"Id": ad_id}
-
-    if ad_type_norm == "TEXT_AD":
-        text_ad = _build_text_ad_update_base(
-            vcard_id,
-            image_hash,
-            sitelink_set_id,
-            callout_setting,
-            clear_image_hash=clear_image_hash,
-        )
-        if title:
-            text_ad["Title"] = title
-        if text:
-            text_ad["Text"] = text
-        if href:
-            text_ad["Href"] = href
-        if title2:
-            text_ad["Title2"] = title2
-        if final_url:
-            text_ad["FinalUrl"] = final_url
-        if display_url_path:
-            text_ad["DisplayUrlPath"] = display_url_path
-        if age_label:
-            text_ad["AgeLabel"] = age_label.upper()
-        if turbo_page_id is not None:
-            text_ad["TurboPageId"] = turbo_page_id
-        if video_extension_creative_id is not None:
-            text_ad["VideoExtension"] = {"CreativeId": video_extension_creative_id}
-        if price_extension:
-            text_ad["PriceExtension"] = price_extension
-        if business_id is not None:
-            text_ad["BusinessId"] = business_id
-        if prefer_vcard_over_business:
-            text_ad["PreferVCardOverBusiness"] = prefer_vcard_over_business.upper()
-        if erir_ad_description:
-            text_ad["ErirAdDescription"] = erir_ad_description
-        if text_ad:
-            ad_data["TextAd"] = text_ad
-    elif ad_type_norm == "DYNAMIC_TEXT_AD":
-        dynamic_text_ad = _build_text_ad_update_base(
-            vcard_id,
-            image_hash,
-            sitelink_set_id,
-            callout_setting,
-            clear_image_hash=clear_image_hash,
-        )
-        if text:
-            dynamic_text_ad["Text"] = text
-        if dynamic_text_ad:
-            ad_data["DynamicTextAd"] = dynamic_text_ad
-    elif ad_type_norm == "TEXT_IMAGE_AD":
-        text_image_ad = {}
-        # No clear_image_hash: ImageAdUpdateBase.AdImageHash is not nillable;
-        # the live API rejects null for this subtype (error 8000).
-        if image_hash:
-            text_image_ad["AdImageHash"] = image_hash
-        if final_url:
-            text_image_ad["FinalUrl"] = final_url
-        if href:
-            text_image_ad["Href"] = href
-        if turbo_page_id is not None:
-            text_image_ad["TurboPageId"] = turbo_page_id
-        if erir_ad_description:
-            text_image_ad["ErirAdDescription"] = erir_ad_description
-        if text_image_ad:
-            ad_data["TextImageAd"] = text_image_ad
-    elif ad_type_norm == "MOBILE_APP_AD":
-        mobile_app_ad = {}
-        if title:
-            mobile_app_ad["Title"] = title
-        if text:
-            mobile_app_ad["Text"] = text
-        if clear_image_hash:
-            mobile_app_ad["AdImageHash"] = None
-        elif image_hash:
-            mobile_app_ad["AdImageHash"] = image_hash
-        if action:
-            mobile_app_ad["Action"] = action.upper()
-        if tracking_url:
-            mobile_app_ad["TrackingUrl"] = tracking_url
-        if parsed_mobile_app_features:
-            mobile_app_ad["Features"] = parsed_mobile_app_features
-        if age_label:
-            mobile_app_ad["AgeLabel"] = age_label.upper()
-        if video_extension_creative_id is not None:
-            mobile_app_ad["VideoExtension"] = {
-                "CreativeId": video_extension_creative_id
-            }
-        if erir_ad_description:
-            mobile_app_ad["ErirAdDescription"] = erir_ad_description
-        if mobile_app_ad:
-            ad_data["MobileAppAd"] = mobile_app_ad
-    elif ad_type_norm == "MOBILE_APP_IMAGE_AD":
-        mobile_app_image_ad = _build_mobile_app_image_ad_update(
-            image_hash,
-            erir_ad_description,
-            tracking_url,
-        )
-        if mobile_app_image_ad:
-            ad_data["MobileAppImageAd"] = mobile_app_image_ad
-    elif ad_type_norm == "RESPONSIVE_AD":
-        responsive_ad = _build_responsive_ad_update(
-            texts,
-            titles,
-            image_hashes,
-            video_extension_ids,
-            sitelink_set_id,
-            callout_setting,
-            href,
-            age_label,
-            display_url_path,
-            price_extension,
-            business_id,
-            erir_ad_description,
-        )
-        if responsive_ad:
-            ad_data["ResponsiveAd"] = responsive_ad
-    elif ad_type_norm in {"SHOPPING_AD", "LISTING_AD"}:
-        feed_based_ad = _build_feed_based_ad_update(
-            sitelink_set_id,
-            callout_setting,
-            business_id,
-            feed_filter_conditions,
-            title_sources,
-            text_sources,
-            default_texts,
-        )
-        if feed_based_ad:
-            field_name = "ShoppingAd" if ad_type_norm == "SHOPPING_AD" else "ListingAd"
-            ad_data[field_name] = feed_based_ad
-    elif ad_type_norm == "SMART_AD_BUILDER_AD":
-        smart_ad_builder_ad = _build_smart_ad_builder_ad_update(
-            logo_extension_hash,
-            erir_ad_description,
-        )
-        if smart_ad_builder_ad:
-            ad_data["SmartAdBuilderAd"] = smart_ad_builder_ad
-    elif ad_type_norm in AD_BUILDER_UPDATE_BLOCKS:
-        ad_builder_ad = _build_ad_builder_update(
-            creative_id,
-            creative_erir_ad_description,
-            erir_ad_description,
-            final_url,
-            href,
-            turbo_page_id,
-            tracking_url,
-            tracking_pixels,
-        )
-        if ad_builder_ad:
-            ad_data[AD_BUILDER_UPDATE_BLOCKS[ad_type_norm]] = ad_builder_ad
-
-    # Reject empty-subtype no-ops: ``{Id: N}`` with no subtype block
-    # is a silent no-op on the live API (issue #198 H1).
-    if len(ad_data) == 1:
-        raise click.UsageError(
-            t(
-                "ads update requires at least one updatable field for --type {ad_type_norm}."
-            ).format(ad_type_norm=ad_type_norm)
-        )
-
-    return ad_data
-
-
-# Documented per-call limit for ads.add is 1000 (Yandex docs, ads/add page);
-# the WSDL declares the Ads array unbounded. ADS_ADD_MAX_BATCH is a conservative
-# CHUNK SIZE (not the ceiling): a partial failure rolls back at most this many
-# ads. 100 keeps round-trips reasonable for hundred-ad uploads while bounding
-# the partial-failure blast radius.
-ADS_ADD_MAX_BATCH = 100
-
-# Batch row keys are the kebab flag names without the leading "--" (so a row
-# reads like `ads add --help`); map them to build_ad_object's dest names.
-_ADS_ROW_KEY_TO_DEST = {label[2:]: dest for dest, label in _ADS_ADD_FLAG_FOR.items()}
-_ADS_ROW_ALLOWED_KEYS = frozenset({"type", "adgroup-id", *_ADS_ROW_KEY_TO_DEST})
-# Multi-value flags accept a JSON list of the existing micro-format strings; keep
-# this in sync with the `multiple=True` add options (--mobile-app-feature,
-# --feed-filter-condition).
-_ADS_ROW_MULTI_KEYS = {"mobile-app-feature", "feed-filter-condition"}
-
-
-def _ads_add_param_types():
-    """Map each ``ads add`` row key (kebab, no ``--``) to its Click ParamType.
-
-    Built lazily from the registered command so a batch row is coerced through
-    the *exact same* type as the single-flag path (issue #562 review): a value
-    like ``--price-extension-price`` (MICRO_RUBLES) or ``--adgroup-id``
-    (IntRange(min=1)) gets the identical conversion/validation, so batch and
-    single produce byte-identical payloads instead of forwarding raw JSON.
-    """
-    types: Dict[str, click.ParamType] = {}
-    for param in add.params:
-        if not isinstance(param, click.Option):
-            continue
-        key = param.opts[0].lstrip("-")
-        # click.STRING is the no-op default; only typed options need coercion.
-        if param.type is not click.STRING:
-            types[key] = param.type
-    return types
-
-
-_ADS_ROW_PARAM_TYPES: Optional[Dict[str, click.ParamType]] = None
-
-
-def _coerce_ad_row_field(key: str, value: Any, row_index: int) -> Any:
-    """Coerce one scalar batch-row value to its single-flag form.
-
-    The CLI only ever sees string tokens, so a batch row must too: this rejects
-    JSON arrays/objects/``null`` for any scalar field, stringifies JSON
-    int/float/bool scalars, then runs typed fields (``param_type is not None``)
-    through their single-flag Click type. The result is that batch and single
-    emit byte-identical payloads — e.g. ``"title": 123`` becomes ``"123"`` (as a
-    CLI token would), ``"adgroup-id": 1.9`` is rejected (not truncated to ``1``),
-    and ``"adgroup-id": null`` / ``[1]`` raise a clear ``Ad row N field`` error
-    instead of an uncaught ``TypeError``. Mirrors keywords' ``_coerce_keyword_field``.
-    """
-    global _ADS_ROW_PARAM_TYPES
-    if _ADS_ROW_PARAM_TYPES is None:
-        _ADS_ROW_PARAM_TYPES = _ads_add_param_types()
-    param_type = _ADS_ROW_PARAM_TYPES.get(key)
-
-    # A scalar flag never accepts a JSON array/object/null — the single-flag CLI
-    # can only express a scalar token. Reject with row/field context up front.
-    if value is None or isinstance(value, (list, dict)):
-        raise click.UsageError(
-            t("Ad row {row_index} field {key!r}: expected a scalar, got {arg0}").format(
-                row_index=row_index, key=key, arg0=type(value).__name__
-            )
-        )
-
-    # The CLI passes every value as a string token; match that so a JSON int for
-    # a string field becomes "123" (not 123) and a typed field parses identically.
-    token = str(value)
-
-    if param_type is None:
-        return token
-
-    if isinstance(value, bool):
-        raise click.UsageError(
-            t("Ad row {row_index} field {key!r}: expected {arg0}, got bool").format(
-                row_index=row_index, key=key, arg0=param_type.name
-            )
-        )
-    try:
-        return param_type.convert(token, None, None)
-    except click.exceptions.BadParameter as exc:
-        raise click.UsageError(
-            t("Ad row {row_index} field {key!r}: {arg0}").format(
-                row_index=row_index, key=key, arg0=exc.format_message()
-            )
-        )
-
-
-def _normalize_ad_row(
-    row: Any, row_index: int, default_adgroup_id: Optional[int]
-) -> Dict[str, Any]:
-    """Translate one flag-form batch row into a built ad object.
-
-    The row keys are kebab flag names without "--" plus ``type`` and
-    ``adgroup-id``. Each typed field is coerced through its single-flag Click
-    type so batch and single emit byte-identical payloads. Unknown keys are
-    rejected; ``build_ad_object`` does the subtype validation, with its
-    UsageError re-raised under an ``Ad row N`` prefix so the operator sees which
-    row failed.
-    """
-    if not isinstance(row, dict):
-        raise click.UsageError(
-            t("Ad row {row_index}: expected JSON object, got {arg0}").format(
-                row_index=row_index, arg0=type(row).__name__
-            )
-        )
-
-    unknown = sorted(set(row) - _ADS_ROW_ALLOWED_KEYS)
-    if unknown:
-        raise click.UsageError(
-            t(
-                "Unknown field {arg0!r} in ad row {row_index}; allowed: {allowed}"
-            ).format(
-                arg0=unknown[0],
-                row_index=row_index,
-                allowed=", ".join(sorted(_ADS_ROW_ALLOWED_KEYS)),
-            )
-        )
-
-    if "adgroup-id" in row:
-        adgroup_id = _coerce_ad_row_field("adgroup-id", row["adgroup-id"], row_index)
-    else:
-        adgroup_id = default_adgroup_id
-    if adgroup_id is None:
-        raise click.UsageError(
-            t(
-                "Ad row {row_index}: missing 'adgroup-id' and no default "
-                "--adgroup-id provided"
-            ).format(row_index=row_index)
-        )
-
-    ad_type = row.get("type")
-    flags: Dict[str, Any] = {}
-    for key, dest in _ADS_ROW_KEY_TO_DEST.items():
-        if key not in row:
-            continue
-        value = row[key]
-        if key in _ADS_ROW_MULTI_KEYS:
-            # A repeatable flag (--mobile-app-feature/--feed-filter-condition) is
-            # a JSON list of the existing micro-format strings; reject anything
-            # else with row/field context instead of crashing downstream.
-            if not isinstance(value, list) or not all(
-                isinstance(item, str) for item in value
-            ):
-                raise click.UsageError(
-                    t(
-                        "Ad row {row_index} field {key!r}: expected a JSON array "
-                        "of strings"
-                    ).format(row_index=row_index, key=key)
-                )
-            value = tuple(value)
-        else:
-            value = _coerce_ad_row_field(key, value, row_index)
-        flags[dest] = value
-
-    mobile_provided = flags.pop("mobile", None)
-
-    try:
-        return build_ad_object(
-            adgroup_id=adgroup_id,
-            ad_type=ad_type,
-            mobile_provided=mobile_provided,
-            flags=flags,
-            flag_for=_ADS_ADD_FLAG_FOR,
-        )
-    except click.UsageError as exc:
-        raise click.UsageError(
-            t("Ad row {row_index}: {arg0}").format(
-                row_index=row_index, arg0=exc.format_message()
-            )
-        )
-
-
-def _bulk_add_ads(ctx, *, adgroup_id, from_file, ads_json, dry_run):
-    if from_file is not None:
-        raw_rows = _batch.load_jsonl_rows(from_file)
-    else:
-        raw_rows = _batch.load_inline_rows(
-            ads_json or "",
-            invalid_json_key="--ads-json: invalid JSON: {arg0}",
-            not_array_key="--ads-json must be a JSON array of ad objects",
-        )
-
-    if not raw_rows:
-        raise click.UsageError(t("Input contains no ad rows."))
-
-    items = [
-        _normalize_ad_row(row, idx, adgroup_id)
-        for idx, row in enumerate(raw_rows, start=1)
-    ]
-
-    _batch.send_batch(
-        ctx,
-        resource="ads",
-        method="add",
-        payload_key="Ads",
-        items=items,
-        max_batch=ADS_ADD_MAX_BATCH,
-        create_client=create_client,
-        dry_run=dry_run,
-        noun="ads",
-    )
-
-
-# Batch row keys for `ads update` are the kebab flag names without "--" plus
-# "id" and "type"; map them to build_ad_update_object's dest names.
-_ADS_UPDATE_ROW_KEY_TO_DEST = {
-    label[2:]: dest for dest, label in _ADS_UPDATE_FLAG_FOR.items()
-}
-_ADS_UPDATE_ROW_ALLOWED_KEYS = frozenset({"id", "type", *_ADS_UPDATE_ROW_KEY_TO_DEST})
-# Repeatable flags accept a JSON list of the existing micro-format strings; keep
-# in sync with the `multiple=True` update options.
-_ADS_UPDATE_ROW_MULTI_KEYS = {"mobile-app-feature", "feed-filter-condition"}
-# --clear-image-hash is a boolean flag (is_flag=True): a row expresses it as a
-# JSON bool, not a string token, so it bypasses _coerce_ad_update_row_field.
-_ADS_UPDATE_ROW_BOOL_KEYS = {"clear-image-hash"}
-
-
-def _ads_update_param_types():
-    """Map each ``ads update`` row key (kebab, no ``--``) to its Click ParamType.
-
-    Built lazily from the registered command so a batch row is coerced through
-    the *exact same* type as the single-flag path (issue #563): e.g.
-    ``--price-extension-price`` (MICRO_RUBLES) or ``--id`` (IntRange(min=1)) get
-    the identical conversion/validation, so batch and single produce
-    byte-identical payloads instead of forwarding raw JSON. Boolean flags
-    (``--clear-image-hash``) are excluded — they are handled as JSON bools.
-    """
-    types: Dict[str, click.ParamType] = {}
-    for param in update.params:
-        if not isinstance(param, click.Option):
-            continue
-        if param.is_flag:
-            continue
-        key = param.opts[0].lstrip("-")
-        # click.STRING is the no-op default; only typed options need coercion.
-        if param.type is not click.STRING:
-            types[key] = param.type
-    return types
-
-
-_ADS_UPDATE_ROW_PARAM_TYPES: Optional[Dict[str, click.ParamType]] = None
-
-
-def _coerce_ad_update_row_field(key: str, value: Any, row_index: int) -> Any:
-    """Coerce one scalar batch-row value to its single-flag form (issue #563).
-
-    Mirrors ``_coerce_ad_row_field``: rejects JSON arrays/objects/``null`` for
-    any scalar field, stringifies JSON int/float/bool scalars, then runs typed
-    fields through their single-flag Click type so batch and single emit
-    byte-identical payloads (``"id": 1.9`` is rejected, not truncated to ``1``;
-    ``"id": null`` / ``[1]`` raise a clear ``Ad update row N field`` error
-    instead of an uncaught ``TypeError``).
-    """
-    global _ADS_UPDATE_ROW_PARAM_TYPES
-    if _ADS_UPDATE_ROW_PARAM_TYPES is None:
-        _ADS_UPDATE_ROW_PARAM_TYPES = _ads_update_param_types()
-    param_type = _ADS_UPDATE_ROW_PARAM_TYPES.get(key)
-
-    if value is None or isinstance(value, (list, dict)):
-        raise click.UsageError(
-            t(
-                "Ad update row {row_index} field {key!r}: expected a scalar, "
-                "got {arg0}"
-            ).format(row_index=row_index, key=key, arg0=type(value).__name__)
-        )
-
-    token = str(value)
-
-    if param_type is None:
-        return token
-
-    if isinstance(value, bool):
-        raise click.UsageError(
-            t(
-                "Ad update row {row_index} field {key!r}: expected {arg0}, got bool"
-            ).format(row_index=row_index, key=key, arg0=param_type.name)
-        )
-    try:
-        return param_type.convert(token, None, None)
-    except click.exceptions.BadParameter as exc:
-        raise click.UsageError(
-            t("Ad update row {row_index} field {key!r}: {arg0}").format(
-                row_index=row_index, key=key, arg0=exc.format_message()
-            )
-        )
-
-
-def _normalize_ad_update_row(row: Any, row_index: int) -> Dict[str, Any]:
-    """Translate one flag-form batch row into a built ad-update object.
-
-    The row keys are kebab flag names without "--" plus ``id`` (required, the
-    update target) and ``type`` (required). Each typed field is coerced through
-    its single-flag Click type so batch and single emit byte-identical payloads.
-    Unknown keys are rejected; ``build_ad_update_object`` does the subtype
-    validation, its UsageError re-raised under an ``Ad update row N`` prefix.
-    """
-    if not isinstance(row, dict):
-        raise click.UsageError(
-            t("Ad update row {row_index}: expected JSON object, got {arg0}").format(
-                row_index=row_index, arg0=type(row).__name__
-            )
-        )
-
-    unknown = sorted(set(row) - _ADS_UPDATE_ROW_ALLOWED_KEYS)
-    if unknown:
-        raise click.UsageError(
-            t(
-                "Unknown field {arg0!r} in ad update row {row_index}; allowed: {allowed}"
-            ).format(
-                arg0=unknown[0],
-                row_index=row_index,
-                allowed=", ".join(sorted(_ADS_UPDATE_ROW_ALLOWED_KEYS)),
-            )
-        )
-
-    if "id" not in row:
-        raise click.UsageError(
-            t("Ad update row {row_index}: missing required 'id'").format(
-                row_index=row_index
-            )
-        )
-    ad_id = _coerce_ad_update_row_field("id", row["id"], row_index)
-
-    ad_type = row.get("type")
-
-    flags: Dict[str, Any] = {}
-    for key, dest in _ADS_UPDATE_ROW_KEY_TO_DEST.items():
-        if key not in row:
-            continue
-        value = row[key]
-        if key in _ADS_UPDATE_ROW_BOOL_KEYS:
-            # A boolean flag (--clear-image-hash) is a JSON bool in a row.
-            if not isinstance(value, bool):
-                raise click.UsageError(
-                    t(
-                        "Ad update row {row_index} field {key!r}: expected a JSON "
-                        "boolean"
-                    ).format(row_index=row_index, key=key)
-                )
-            # clear-image-hash:false is the flag-absent state (skip it); only a
-            # true value sets the flag, matching the single-flag command.
-            if not value:
-                continue
-        elif key in _ADS_UPDATE_ROW_MULTI_KEYS:
-            if not isinstance(value, list) or not all(
-                isinstance(item, str) for item in value
-            ):
-                raise click.UsageError(
-                    t(
-                        "Ad update row {row_index} field {key!r}: expected a JSON "
-                        "array of strings"
-                    ).format(row_index=row_index, key=key)
-                )
-            value = tuple(value)
-        else:
-            value = _coerce_ad_update_row_field(key, value, row_index)
-        flags[dest] = value
-
-    # Match single-mode ordering: the command validates --type before the
-    # --image-hash/--clear-image-hash mutex, so check the missing 'type' first.
-    if ad_type is None:
-        raise click.UsageError(
-            t("Ad update row {row_index}: missing required 'type'").format(
-                row_index=row_index
-            )
-        )
-
-    # Reproduce the command-level --image-hash/--clear-image-hash mutex per row.
-    if flags.get("image_hash") and flags.get("clear_image_hash"):
-        raise click.UsageError(
-            t(
-                "Ad update row {row_index}: use either 'image-hash' or "
-                "'clear-image-hash', not both"
-            ).format(row_index=row_index)
-        )
-
-    try:
-        return build_ad_update_object(
-            ad_id=ad_id,
-            ad_type=ad_type,
-            flags=flags,
-            flag_for=_ADS_UPDATE_FLAG_FOR,
-        )
-    except click.UsageError as exc:
-        raise click.UsageError(
-            t("Ad update row {row_index}: {arg0}").format(
-                row_index=row_index, arg0=exc.format_message()
-            )
-        )
-
-
-def _bulk_update_ads(ctx, *, from_file, ads_json, dry_run):
-    if from_file is not None:
-        raw_rows = _batch.load_jsonl_rows(from_file)
-    else:
-        raw_rows = _batch.load_inline_rows(
-            ads_json or "",
-            invalid_json_key="--ads-json: invalid JSON: {arg0}",
-            not_array_key="--ads-json must be a JSON array of ad objects",
-        )
-
-    if not raw_rows:
-        raise click.UsageError(t("Input contains no ad rows."))
-
-    items = [
-        _normalize_ad_update_row(row, idx) for idx, row in enumerate(raw_rows, start=1)
-    ]
-
-    _batch.send_batch(
-        ctx,
-        resource="ads",
-        method="update",
-        payload_key="Ads",
-        items=items,
-        max_batch=ADS_ADD_MAX_BATCH,
-        create_client=create_client,
-        dry_run=dry_run,
-        noun="ads",
-        result_key="UpdateResults",
-    )
-
-
-@ads.command()
-@click.option(
-    "--adgroup-id",
-    type=click.IntRange(min=1),
-    help="Ad group ID (required in single-item mode; batch default in --from-file mode)",
-)
-@click.option(
-    "--type",
-    "ad_type",
-    default="TEXT_AD",
-    help=(
-        "Ad type: TEXT_AD | TEXT_IMAGE_AD | MOBILE_APP_AD | DYNAMIC_TEXT_AD | "
-        "MOBILE_APP_IMAGE_AD | RESPONSIVE_AD | SHOPPING_AD | LISTING_AD | "
-        "SMART_AD_BUILDER_AD | TEXT_AD_BUILDER_AD | "
-        "MOBILE_APP_AD_BUILDER_AD | MOBILE_APP_CPC_VIDEO_AD_BUILDER_AD | "
-        "CPC_VIDEO_AD_BUILDER_AD | CPM_BANNER_AD_BUILDER_AD | "
-        "CPM_VIDEO_AD_BUILDER_AD"
-    ),
-)
-@click.option("--title", help="Ad title (TEXT_AD / MOBILE_APP_AD)")
-@click.option("--text", help="Ad text (TEXT_AD / MOBILE_APP_AD / DYNAMIC_TEXT_AD)")
-@click.option("--titles", help="Comma-separated ResponsiveAd.Titles values")
-@click.option("--texts", help="Comma-separated ResponsiveAd.Texts values")
-@click.option(
-    "--href",
-    help=(
-        "Ad URL (TEXT_AD / TEXT_IMAGE_AD / RESPONSIVE_AD / TEXT_AD_BUILDER_AD / "
-        "CPC_VIDEO_AD_BUILDER_AD / CPM_BANNER_AD_BUILDER_AD / "
-        "CPM_VIDEO_AD_BUILDER_AD)"
-    ),
-)
-@click.option(
-    "--image-hash",
-    help=(
-        "Ad image hash (TEXT_AD / TEXT_IMAGE_AD / MOBILE_APP_AD / "
-        "DYNAMIC_TEXT_AD / MOBILE_APP_IMAGE_AD)"
-    ),
-)
-@click.option(
-    "--image-hashes",
-    help="Comma-separated ResponsiveAd.AdImageHashes values",
-)
-@click.option(
-    "--action",
-    help="MOBILE_APP_AD call-to-action (MobileAppAdActionEnum, e.g. INSTALL)",
-)
-@click.option(
-    "--tracking-url",
-    help=(
-        "Tracking URL (MOBILE_APP_AD / MOBILE_APP_AD_BUILDER_AD / "
-        "MOBILE_APP_CPC_VIDEO_AD_BUILDER_AD / MOBILE_APP_IMAGE_AD)"
-    ),
-)
-@click.option(
-    "--age-label",
-    help="Age label (MOBILE_APP_AD MobAppAgeLabelEnum / RESPONSIVE_AD AgeLabelEnum)",
-)
-@click.option(
-    "--mobile-app-feature",
-    "mobile_app_features",
-    multiple=True,
-    help=(
-        "Repeatable MobileAppAd.Features item as FEATURE=YES|NO "
-        "(PRICE, ICON, CUSTOMER_RATING, RATINGS)"
-    ),
-)
-@click.option("--title2", help="Second headline (TEXT_AD)")
-@click.option("--display-url-path", help="Display URL path (TEXT_AD / RESPONSIVE_AD)")
-@click.option(
-    "--mobile",
-    type=click.Choice(["YES", "NO"], case_sensitive=False),
-    default="NO",
-    help="Mobile-targeted flag (TEXT_AD)",
-)
-@click.option("--vcard-id", type=int, help="VCard ID (TEXT_AD / DYNAMIC_TEXT_AD)")
-@click.option(
-    "--sitelink-set-id",
-    type=int,
-    help=(
-        "Sitelink set ID "
-        "(TEXT_AD / DYNAMIC_TEXT_AD / RESPONSIVE_AD / SHOPPING_AD / LISTING_AD)"
-    ),
-)
-@click.option(
-    "--turbo-page-id",
-    type=int,
-    help=(
-        "Turbo page ID (TEXT_AD / TEXT_IMAGE_AD / TEXT_AD_BUILDER_AD / "
-        "CPC_VIDEO_AD_BUILDER_AD / CPM_BANNER_AD_BUILDER_AD / "
-        "CPM_VIDEO_AD_BUILDER_AD)"
-    ),
-)
-@click.option(
-    "--ad-extensions",
-    help=(
-        "Comma-separated ad extension IDs "
-        "(TEXT_AD / DYNAMIC_TEXT_AD / RESPONSIVE_AD / SHOPPING_AD / LISTING_AD)"
-    ),
-)
-@click.option(
-    "--final-url",
-    help="FinalUrl (TEXT_AD / TEXT_IMAGE_AD / TEXT_AD_BUILDER_AD)",
-)
-@click.option(
-    "--video-extension-creative-id",
-    type=int,
-    help="TextAd/MobileAppAd.VideoExtension.CreativeId (TEXT_AD / MOBILE_APP_AD)",
-)
-@click.option(
-    "--price-extension-price",
-    type=MICRO_RUBLES,
-    help=(
-        "TextAd/ResponsiveAd.PriceExtension.Price in micro-rubles. "
-        "Required whenever any PriceExtension flag is used."
-    ),
-)
-@click.option(
-    "--price-extension-old-price",
-    type=MICRO_RUBLES,
-    help=(
-        "TextAd/ResponsiveAd.PriceExtension.OldPrice in micro-rubles. "
-        "Optional; if supplied, PriceExtension add also requires "
-        "--price-extension-price, --price-extension-price-qualifier, "
-        "and --price-extension-price-currency."
-    ),
-)
-@click.option(
-    "--price-extension-price-qualifier",
-    type=click.Choice(["FROM", "UP_TO", "NONE"], case_sensitive=False),
-    help=(
-        "TextAd/ResponsiveAd.PriceExtension.PriceQualifier: FROM, UP_TO, or NONE. "
-        "Required whenever any PriceExtension flag is used."
-    ),
-)
-@click.option(
-    "--price-extension-price-currency",
-    type=click.Choice(
-        ["RUB", "UAH", "BYN", "USD", "EUR", "KZT", "TRY", "CHF", "UZS"],
-        case_sensitive=False,
-    ),
-    help=(
-        "TextAd/ResponsiveAd.PriceExtension.PriceCurrency enum value. "
-        "Required whenever any PriceExtension flag is used."
-    ),
-)
-@click.option(
-    "--video-extension-ids",
-    help="Comma-separated ResponsiveAd.VideoExtensionIds values",
-)
-@click.option(
-    "--business-id",
-    type=int,
-    help="BusinessId (TEXT_AD / RESPONSIVE_AD / SHOPPING_AD / LISTING_AD)",
-)
-@click.option(
-    "--prefer-vcard-over-business",
-    type=click.Choice(["YES", "NO"], case_sensitive=False),
-    help="TextAd.PreferVCardOverBusiness value: YES or NO",
-)
-@click.option(
-    "--erir-ad-description",
-    help=(
-        "ErirAdDescription (TEXT_AD / TEXT_IMAGE_AD / MOBILE_APP_AD / "
-        "MOBILE_APP_IMAGE_AD / RESPONSIVE_AD / non-SMART AdBuilder add subtypes)"
-    ),
-)
-@click.option(
-    "--creative-id",
-    type=int,
-    help="AdBuilder Creative.CreativeId for non-SMART AdBuilder add subtypes",
-)
-@click.option(
-    "--tracking-pixels",
-    help="Comma-separated AdBuilder TrackingPixels.Items values",
-)
-@click.option(
-    "--logo-extension-hash",
-    help="SmartAdBuilderAd.LogoExtensionHash (SMART_AD_BUILDER_AD)",
-)
-@click.option(
-    "--feed-id",
-    type=int,
-    help="ShoppingAd/ListingAd.FeedId (SHOPPING_AD / LISTING_AD)",
-)
-@click.option(
-    "--feed-filter-condition",
-    "feed_filter_conditions",
-    multiple=True,
-    help=(
-        "Repeatable ShoppingAd/ListingAd.FeedFilterConditions item as "
-        "OPERAND:OPERATOR:ARG1|ARG2"
-    ),
-)
-@click.option(
-    "--title-sources",
-    help="Comma-separated ShoppingAd/ListingAd.TitleSources values",
-)
-@click.option(
-    "--text-sources",
-    help="Comma-separated ShoppingAd/ListingAd.TextSources values",
-)
-@click.option(
-    "--default-texts",
-    help="ShoppingAd/ListingAd.DefaultTexts value (required for SHOPPING_AD/LISTING_AD)",
-)
-@click.option(
-    "--from-file",
-    "from_file",
-    type=click.Path(exists=True, dir_okay=False, readable=True),
-    help="Path to a JSONL file (one flag-form ad object per line) for batch add",
-)
-@click.option(
-    "--ads-json",
-    "ads_json",
-    help="Inline JSON array of flag-form ad objects for batch add",
-)
-@click.option("--dry-run", is_flag=True, help="Show request without sending")
-@click.pass_context
-@handle_api_errors
-def add(
-    ctx,
-    adgroup_id,
-    ad_type,
-    title,
-    text,
-    titles,
-    texts,
-    href,
-    image_hash,
-    image_hashes,
-    action,
-    tracking_url,
-    age_label,
-    mobile_app_features,
-    title2,
-    display_url_path,
-    mobile,
-    vcard_id,
-    sitelink_set_id,
-    turbo_page_id,
-    ad_extensions,
-    final_url,
-    video_extension_creative_id,
-    price_extension_price,
-    price_extension_old_price,
-    price_extension_price_qualifier,
-    price_extension_price_currency,
-    video_extension_ids,
-    business_id,
-    prefer_vcard_over_business,
-    erir_ad_description,
-    creative_id,
-    tracking_pixels,
-    logo_extension_hash,
-    feed_id,
-    feed_filter_conditions,
-    title_sources,
-    text_sources,
-    default_texts,
-    from_file,
-    ads_json,
-    dry_run,
-):
-    """Add one or many ads.
-
-    Single-item mode uses typed flags (--type, --title, ...). Batch mode reads
-    flag-form rows from --from-file (JSONL, one object per line) or --ads-json
-    (inline JSON array); each row is the same flag set keyed by the kebab flag
-    name without the leading dashes (e.g. {"type":"TEXT_AD","title":"...",
-    "text":"...","href":"...","adgroup-id":1}). --adgroup-id is the batch
-    default and may be overridden per row.
-    """
-    flags_local = {
-        "title": title,
-        "text": text,
-        "titles": titles,
-        "texts": texts,
-        "href": href,
-        "image_hash": image_hash,
-        "image_hashes": image_hashes,
-        "action": action,
-        "tracking_url": tracking_url,
-        "age_label": age_label,
-        "mobile_app_features": mobile_app_features,
-        "title2": title2,
-        "display_url_path": display_url_path,
-        "vcard_id": vcard_id,
-        "sitelink_set_id": sitelink_set_id,
-        "turbo_page_id": turbo_page_id,
-        "ad_extensions": ad_extensions,
-        "final_url": final_url,
-        "video_extension_creative_id": video_extension_creative_id,
-        "price_extension_price": price_extension_price,
-        "price_extension_old_price": price_extension_old_price,
-        "price_extension_price_qualifier": price_extension_price_qualifier,
-        "price_extension_price_currency": price_extension_price_currency,
-        "video_extension_ids": video_extension_ids,
-        "business_id": business_id,
-        "prefer_vcard_over_business": prefer_vcard_over_business,
-        "erir_ad_description": erir_ad_description,
-        "creative_id": creative_id,
-        "tracking_pixels": tracking_pixels,
-        "logo_extension_hash": logo_extension_hash,
-        "feed_id": feed_id,
-        "feed_filter_conditions": feed_filter_conditions,
-        "title_sources": title_sources,
-        "text_sources": text_sources,
-        "default_texts": default_texts,
-    }
-
-    modes_used = sum(1 for v in (from_file, ads_json) if v is not None)
-    if modes_used > 1:
-        raise click.UsageError(
-            t(
-                "Provide at most one of: --from-file or --ads-json \u2014 "
-                "they are mutually exclusive."
-            )
-        )
-    batch_mode = modes_used > 0
-
-    mobile_source = ctx.get_parameter_source("mobile")
-    mobile_explicit = (
-        mobile_source != click.core.ParameterSource.DEFAULT if mobile_source else False
-    )
-    mobile_provided = mobile if mobile_explicit else None
-
-    if batch_mode:
-        batch_incompatible = {
-            label: (mobile_provided if dest == "mobile" else flags_local.get(dest))
-            for dest, label in _ADS_ADD_FLAG_FOR.items()
-        }
-        # --type carries a Click default ("TEXT_AD"); only count it as provided
-        # when the operator actually passed it.
-        type_source = ctx.get_parameter_source("ad_type")
-        type_explicit = (
-            type_source != click.core.ParameterSource.DEFAULT if type_source else False
-        )
-        if type_explicit:
-            batch_incompatible["--type"] = ad_type
-        unsupported = sorted(
-            label
-            for label, value in batch_incompatible.items()
-            if value not in (None, ())
-        )
-        if unsupported:
-            raise click.UsageError(
-                t("{arg0} supported only with single-item mode").format(
-                    arg0=", ".join(unsupported)
-                )
-            )
-        _bulk_add_ads(
-            ctx,
-            adgroup_id=adgroup_id,
-            from_file=from_file,
-            ads_json=ads_json,
-            dry_run=dry_run,
-        )
-        return
-
-    if adgroup_id is None:
-        raise click.UsageError(t("Missing option '--adgroup-id'."))
-
-    ad_data = build_ad_object(
-        adgroup_id=adgroup_id,
-        ad_type=ad_type,
-        mobile_provided=mobile_provided,
-        flags=flags_local,
-        flag_for=_ADS_ADD_FLAG_FOR,
-    )
-
-    body = {"method": "add", "params": {"Ads": [ad_data]}}
-
-    execute_request(ctx, "ads", body, dry_run, create_client)
-
-
-@ads.command()
-@click.option(
-    "--id",
-    "ad_id",
-    type=click.IntRange(min=1),
-    help="Ad ID (required in single-item mode; per row in --from-file mode)",
-)
-@click.option(
-    "--type",
-    "ad_type",
-    help=(
-        "Ad subtype: TEXT_AD | TEXT_IMAGE_AD | MOBILE_APP_AD | "
-        "DYNAMIC_TEXT_AD | MOBILE_APP_IMAGE_AD | RESPONSIVE_AD | "
-        "SHOPPING_AD | LISTING_AD | SMART_AD_BUILDER_AD | TEXT_AD_BUILDER_AD | "
-        "MOBILE_APP_AD_BUILDER_AD | MOBILE_APP_CPC_VIDEO_AD_BUILDER_AD | "
-        "CPC_VIDEO_AD_BUILDER_AD | CPM_BANNER_AD_BUILDER_AD | "
-        "CPM_VIDEO_AD_BUILDER_AD "
-        "(required in single-item mode; per row in --from-file mode)"
-    ),
-)
-@click.option(
-    "--status",
-    help=(
-        "Deprecated: not part of WSDL AdUpdateItem. "
-        "Use 'direct ads suspend/resume/archive/unarchive' instead."
-    ),
-)
-@click.option("--title", help="Title (TEXT_AD / MOBILE_APP_AD)")
-@click.option("--text", help="Text (TEXT_AD / MOBILE_APP_AD / DYNAMIC_TEXT_AD)")
-@click.option("--titles", help="Comma-separated ResponsiveAd.Titles values")
-@click.option("--texts", help="Comma-separated ResponsiveAd.Texts values")
-@click.option(
-    "--href",
-    help=(
-        "URL (TEXT_AD / TEXT_IMAGE_AD / RESPONSIVE_AD / "
-        "TEXT_AD_BUILDER_AD / CPC_VIDEO_AD_BUILDER_AD / "
-        "CPM_BANNER_AD_BUILDER_AD / CPM_VIDEO_AD_BUILDER_AD)"
-    ),
-)
-@click.option(
-    "--image-hash",
-    help=(
-        "Image hash (TEXT_AD / TEXT_IMAGE_AD / MOBILE_APP_AD / "
-        "DYNAMIC_TEXT_AD / MOBILE_APP_IMAGE_AD)"
-    ),
-)
-@click.option(
-    "--clear-image-hash",
-    is_flag=True,
-    help=(
-        "Set AdImageHash to null to remove the image (TEXT_AD / DYNAMIC_TEXT_AD / "
-        "MOBILE_APP_AD). Not available for TEXT_IMAGE_AD / MOBILE_APP_IMAGE_AD: "
-        "their AdImageHash is not nillable and the API rejects null. "
-        "Note: if the ad has a server-side carousel image, Yandex rejects the "
-        "reset with Error 5005 (carousel images are not exposed by Ads.update "
-        "and can only be removed in the Direct web interface); replace it with a "
-        "different --image-hash instead."
-    ),
-)
-@click.option(
-    "--image-hashes",
-    help="Comma-separated ResponsiveAd.AdImageHashes.Items values",
-)
-@click.option(
-    "--action",
-    help="MOBILE_APP_AD call-to-action (MobileAppAdActionEnum, e.g. INSTALL)",
-)
-@click.option(
-    "--tracking-url",
-    help=(
-        "Tracking URL (MOBILE_APP_AD / MOBILE_APP_AD_BUILDER_AD / "
-        "MOBILE_APP_CPC_VIDEO_AD_BUILDER_AD / MOBILE_APP_IMAGE_AD)"
-    ),
-)
-@click.option(
-    "--age-label",
-    help=(
-        "Age label (TEXT_AD / RESPONSIVE_AD AgeLabelEnum; "
-        "MOBILE_APP_AD MobAppAgeLabelEnum)"
-    ),
-)
-@click.option(
-    "--mobile-app-feature",
-    "mobile_app_features",
-    multiple=True,
-    help=(
-        "Repeatable MobileAppAd.Features item as FEATURE=YES|NO. "
-        "Features: PRICE, ICON, CUSTOMER_RATING, RATINGS."
-    ),
-)
-@click.option("--title2", help="Second headline (TEXT_AD)")
-@click.option("--display-url-path", help="Display URL path (TEXT_AD / RESPONSIVE_AD)")
-@click.option("--vcard-id", type=int, help="VCard ID (TEXT_AD / DYNAMIC_TEXT_AD)")
-@click.option(
-    "--sitelink-set-id",
-    type=int,
-    help=(
-        "Sitelink set ID "
-        "(TEXT_AD / DYNAMIC_TEXT_AD / RESPONSIVE_AD / SHOPPING_AD / LISTING_AD)"
-    ),
-)
-@click.option(
-    "--turbo-page-id",
-    type=int,
-    help=(
-        "Turbo page ID (TEXT_AD / TEXT_IMAGE_AD / TEXT_AD_BUILDER_AD / "
-        "CPC_VIDEO_AD_BUILDER_AD / CPM_BANNER_AD_BUILDER_AD / "
-        "CPM_VIDEO_AD_BUILDER_AD)"
-    ),
-)
-@click.option(
-    "--callouts-add",
-    help=(
-        "Comma-separated CALLOUT ad-extension IDs to attach "
-        "(Operation=ADD). TEXT_AD / DYNAMIC_TEXT_AD / RESPONSIVE_AD / "
-        "SHOPPING_AD / LISTING_AD only."
-    ),
-)
-@click.option(
-    "--callouts-remove",
-    help=(
-        "Comma-separated CALLOUT ad-extension IDs to detach "
-        "(Operation=REMOVE). TEXT_AD / DYNAMIC_TEXT_AD / RESPONSIVE_AD / "
-        "SHOPPING_AD / LISTING_AD only."
-    ),
-)
-@click.option(
-    "--callouts-set",
-    help=(
-        "Comma-separated CALLOUT ad-extension IDs that REPLACE the ad's "
-        "current callout list (Operation=SET). Mutually exclusive with "
-        "--callouts-add / --callouts-remove. TEXT_AD / DYNAMIC_TEXT_AD / "
-        "RESPONSIVE_AD / SHOPPING_AD / LISTING_AD only."
-    ),
-)
-@click.option(
-    "--video-extension-creative-id",
-    type=int,
-    help=(
-        "Video extension CreativeId for TextAd/MobileAppAd.VideoExtension. "
-        "TEXT_AD / MOBILE_APP_AD only."
-    ),
-)
-@click.option(
-    "--video-extension-ids",
-    help="Comma-separated ResponsiveAd.VideoExtensionIds.Items values",
-)
-@click.option(
-    "--price-extension-price",
-    type=MICRO_RUBLES,
-    help=("PriceExtension.Price in micro-rubles. TEXT_AD / RESPONSIVE_AD only."),
-)
-@click.option(
-    "--price-extension-old-price",
-    type=MICRO_RUBLES,
-    help=("PriceExtension.OldPrice in micro-rubles. TEXT_AD / RESPONSIVE_AD only."),
-)
-@click.option(
-    "--price-extension-price-qualifier",
-    type=click.Choice(["FROM", "UP_TO", "NONE"], case_sensitive=False),
-    help=(
-        "PriceExtension.PriceQualifier: FROM, UP_TO, or NONE. "
-        "TEXT_AD / RESPONSIVE_AD only."
-    ),
-)
-@click.option(
-    "--price-extension-price-currency",
-    type=click.Choice(
-        ["RUB", "UAH", "BYN", "USD", "EUR", "KZT", "TRY", "CHF", "UZS"],
-        case_sensitive=False,
-    ),
-    help="PriceExtension.PriceCurrency enum value. TEXT_AD / RESPONSIVE_AD only.",
-)
-@click.option(
-    "--business-id",
-    type=int,
-    help="BusinessId (TEXT_AD / RESPONSIVE_AD / SHOPPING_AD / LISTING_AD)",
-)
-@click.option(
-    "--prefer-vcard-over-business",
-    type=click.Choice(["YES", "NO"], case_sensitive=False),
-    help="TextAd.PreferVCardOverBusiness value: YES or NO",
-)
-@click.option(
-    "--erir-ad-description",
-    help=(
-        "ErirAdDescription for TEXT_AD, TEXT_IMAGE_AD, MOBILE_APP_AD, "
-        "RESPONSIVE_AD, MOBILE_APP_IMAGE_AD, SMART_AD_BUILDER_AD, "
-        "and AdBuilder update subtypes"
-    ),
-)
-@click.option(
-    "--logo-extension-hash",
-    help="SmartAdBuilderAd.LogoExtensionHash",
-)
-@click.option(
-    "--creative-id",
-    type=int,
-    help="AdBuilder Creative.CreativeId for AdBuilder update subtypes",
-)
-@click.option(
-    "--creative-erir-ad-description",
-    help="AdBuilder Creative.ErirAdDescription; requires --creative-id",
-)
-@click.option(
-    "--final-url",
-    help="FinalUrl (TEXT_AD / TEXT_IMAGE_AD / TEXT_AD_BUILDER_AD)",
-)
-@click.option(
-    "--tracking-pixels",
-    help="Comma-separated AdBuilder TrackingPixels.Items values",
-)
-@click.option(
-    "--feed-filter-condition",
-    "feed_filter_conditions",
-    multiple=True,
-    help=(
-        "Repeatable ShoppingAd/ListingAd FeedFilterConditions item as "
-        "OPERAND:OPERATOR:ARG1|ARG2"
-    ),
-)
-@click.option(
-    "--title-sources",
-    help="Comma-separated ShoppingAd/ListingAd.TitleSources.Items values",
-)
-@click.option(
-    "--text-sources",
-    help="Comma-separated ShoppingAd/ListingAd.TextSources.Items values",
-)
-@click.option(
-    "--default-texts",
-    help="Comma-separated ShoppingAd/ListingAd.DefaultTexts values",
-)
-@click.option(
-    "--from-file",
-    "from_file",
-    type=click.Path(exists=True, dir_okay=False, readable=True),
-    help="Path to a JSONL file (one flag-form ad-update object per line) for batch update",
-)
-@click.option(
-    "--ads-json",
-    "ads_json",
-    help="Inline JSON array of flag-form ad-update objects for batch update",
-)
-@click.option("--dry-run", is_flag=True, help="Show request without sending")
-@click.pass_context
-@handle_api_errors
-def update(
-    ctx,
-    ad_id,
-    ad_type,
-    status,
-    title,
-    text,
-    titles,
-    texts,
-    href,
-    image_hash,
-    clear_image_hash,
-    image_hashes,
-    action,
-    tracking_url,
-    age_label,
-    mobile_app_features,
-    title2,
-    display_url_path,
-    vcard_id,
-    sitelink_set_id,
-    turbo_page_id,
-    callouts_add,
-    callouts_remove,
-    callouts_set,
-    video_extension_creative_id,
-    video_extension_ids,
-    price_extension_price,
-    price_extension_old_price,
-    price_extension_price_qualifier,
-    price_extension_price_currency,
-    business_id,
-    prefer_vcard_over_business,
-    erir_ad_description,
-    logo_extension_hash,
-    creative_id,
-    creative_erir_ad_description,
-    final_url,
-    tracking_pixels,
-    feed_filter_conditions,
-    title_sources,
-    text_sources,
-    default_texts,
-    from_file,
-    ads_json,
-    dry_run,
-):
-    """Update one or many ads.
-
-    Single-item mode uses typed flags (--id, --type, ...). Batch mode reads
-    flag-form rows from --from-file (JSONL, one object per line) or --ads-json
-    (inline JSON array); each row is the same flag set keyed by the kebab flag
-    name without the leading dashes (e.g. {"id":5,"type":"TEXT_AD",
-    "title":"New"}). Each row carries its own "id".
-    """
-    if status:
-        raise click.UsageError(
-            t(
-                "Use 'direct ads suspend/resume/archive/unarchive' to change status. "
-                "The --status flag is not supported by WSDL AdUpdateItem."
-            )
-        )
-
-    flags_local = {
-        "title": title,
-        "text": text,
-        "titles": titles,
-        "texts": texts,
-        "href": href,
-        "image_hash": image_hash,
-        "clear_image_hash": clear_image_hash or None,
-        "image_hashes": image_hashes,
-        "action": action,
-        "tracking_url": tracking_url,
-        "age_label": age_label,
-        "mobile_app_features": mobile_app_features,
-        "title2": title2,
-        "display_url_path": display_url_path,
-        "vcard_id": vcard_id,
-        "sitelink_set_id": sitelink_set_id,
-        "turbo_page_id": turbo_page_id,
-        "callouts_add": callouts_add,
-        "callouts_remove": callouts_remove,
-        "callouts_set": callouts_set,
-        "video_extension_creative_id": video_extension_creative_id,
-        "video_extension_ids": video_extension_ids,
-        "price_extension_price": price_extension_price,
-        "price_extension_old_price": price_extension_old_price,
-        "price_extension_price_qualifier": price_extension_price_qualifier,
-        "price_extension_price_currency": price_extension_price_currency,
-        "business_id": business_id,
-        "prefer_vcard_over_business": prefer_vcard_over_business,
-        "erir_ad_description": erir_ad_description,
-        "logo_extension_hash": logo_extension_hash,
-        "creative_id": creative_id,
-        "creative_erir_ad_description": creative_erir_ad_description,
-        "final_url": final_url,
-        "tracking_pixels": tracking_pixels,
-        "feed_filter_conditions": feed_filter_conditions,
-        "title_sources": title_sources,
-        "text_sources": text_sources,
-        "default_texts": default_texts,
-    }
-
-    modes_used = sum(1 for v in (from_file, ads_json) if v is not None)
-    if modes_used > 1:
-        raise click.UsageError(
-            t(
-                "Provide at most one of: --from-file or --ads-json — "
-                "they are mutually exclusive."
-            )
-        )
-    batch_mode = modes_used > 0
-
-    if batch_mode:
-        batch_incompatible = {
-            label: flags_local.get(dest) for dest, label in _ADS_UPDATE_FLAG_FOR.items()
-        }
-        # --type/--id have no Click default, so an absent flag is None; in batch
-        # mode the operator passes them per row, not on the command line.
-        if ad_type is not None:
-            batch_incompatible["--type"] = ad_type
-        if ad_id is not None:
-            batch_incompatible["--id"] = ad_id
-        unsupported = sorted(
-            label
-            for label, value in batch_incompatible.items()
-            if value not in (None, ())
-        )
-        if unsupported:
-            raise click.UsageError(
-                t("{arg0} supported only with single-item mode").format(
-                    arg0=", ".join(unsupported)
-                )
-            )
-        _bulk_update_ads(
-            ctx,
-            from_file=from_file,
-            ads_json=ads_json,
-            dry_run=dry_run,
-        )
-        return
-
-    if ad_id is None:
-        raise click.UsageError(t("Missing option '--id'."))
-
-    if ad_type is None:
-        raise click.UsageError(t("Missing option '--type'."))
-
-    if image_hash and clear_image_hash:
-        raise click.UsageError(
-            t("Use either --image-hash or --clear-image-hash, not both")
-        )
-
-    ad_data = build_ad_update_object(
-        ad_id=ad_id,
-        ad_type=ad_type,
-        flags=flags_local,
-        flag_for=_ADS_UPDATE_FLAG_FOR,
-    )
-
-    body = {"method": "update", "params": {"Ads": [ad_data]}}
-
-    execute_request(ctx, "ads", body, dry_run, create_client)
-
-
-register_lifecycle_commands(
-    ads,
-    "ad_id",
-    "Ad ID",
-    create_client,
-    [
-        ("delete", "Delete ad"),
-        ("archive", "Archive ad"),
-        ("unarchive", "Unarchive ad"),
-        ("suspend", "Suspend ad"),
-        ("resume", "Resume ad"),
-        ("moderate", "Moderate ad"),
-    ],
-)
